Introduce MessagePump to represent the native message pump used to drive a MessageLoop.  A MessageLoop now has a MessagePump.
This will make it possible to port the MessagePump interface to other platforms
as well as to use an IO completion port for our worker threads on Windows.

Currently, there is only MessagePumpWin, which attempts to preserve the
pre-existing behavior of the MessageLoop.

API changes to MessageLoop:

1.  MessageLoop::Quit means return from Run when the MessageLoop would
otherwise wait for more work.

2.  MessageLoop::Quit can no longer be called outside the context of an active Run
call.  So, things like this:
  MessageLoop::current()->Quit();
  MessageLoop::current()->Run();
are now:
  MessageLoop::current()->RunAllPending();

3.  MessageLoop::Quit can no longer be called from other threads.  This means that
PostTask(..., new MessageLoop::QuitTask()) must be used explicitly to Quit across
thread boundaries.

4.  No protection is made to deal with nested MessageLoops involving watched
objects or APCs.  In fact, an assertion is added to flag such cases.  This is a
temporary measure until object watching and APC facilities are removed in favor
of a MessagePump designed around an IO completion port.

As part of this CL, I also changed the automation system to use an
IPC::ChannelProxy instead of an IPC::Channel.  This moves the automation IPC
onto Chrome's IO thread where it belongs.  I also fixed some abuses of
RefCounted in the AutomationProvider class.  It was deleting itself in some
cases!  This led to having to fix the ownership model for AutomationProvider,
which explains the changes to AutomationProviderList and so on.

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@928 0039d316-1c4b-4281-b951-d872f2087c98


CrOS-Libchrome-Original-Commit: 295039bdf97f08bf5c1c1c136dd977b7e97ddd31