Reverting 7423.
After talking with Darin, I'm convinced this was a bad idea.  The issue is that Linux can't actually know the encoding of the file (bad Linux!), so on Linux this method is nonsensical.  It also turns out to not be useful for the case I thought I needed it for.
Review URL: http://codereview.chromium.org/16252

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@7455 0039d316-1c4b-4281-b951-d872f2087c98


CrOS-Libchrome-Original-Commit: a38b8719287643f591d4539b43f661008d28bbfd