Remove some unused stuff from basictypes.h.
My original goal was just to remove down_cast, since it uses RTTI, and I am
going to update the build with -fno-rtti.  The other stuff I just noticed is
unrelated and unused in our code, so we might as well remove it, since it
hasn't proven to be useful.

Review URL: http://codereview.chromium.org/63121


git-svn-id: svn://svn.chromium.org/chrome/trunk/src@13349 0039d316-1c4b-4281-b951-d872f2087c98


CrOS-Libchrome-Original-Commit: 7340a68ff7f45460c0a71f1397737079cbf4d59d