Linux: add internal core dump reporting.
In order to get a handle on a number of browser crashes, this patch
adds an options for Google internal folks to upload full core files
when the browser crashes. The core files are uploaded to an internal
IP address and will be available to other developers.

This is only enabled for Chrome branded builds and then, only if
--google-internal-crash-reporting is given on the command line.

http://codereview.chromium.org/160212


git-svn-id: svn://svn.chromium.org/chrome/trunk/src@21782 0039d316-1c4b-4281-b951-d872f2087c98


CrOS-Libchrome-Original-Commit: abaa71c990b2d9dd514e9f2eefb1d2cf7b720562