linux: unload plugin libraries after extracting mime types
I had not unloaded them as a premature optimization, thinking it'd be
better to leave 'em around since they'll eventually be loaded by a page.
But that happens in the plugin process and reading the mime type happens
in the browser process, so it didn't help anyway.

BUG=17444

Review URL: http://codereview.chromium.org/160303

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@21929 0039d316-1c4b-4281-b951-d872f2087c98


CrOS-Libchrome-Original-Commit: 803ef4ef5b14ff135b826425825291259f9ceeab