Use 'icu::' namespace explicitly throughout Chrome tree instead of relying on 'using namespace icu'.
This is Chrome's counterpart to the ICU header change that disables
'using namespace icu' (http://codereview.chromium.org/171010/show), which is required to avoid the name colission between Chrome's StringPiece (in base) and ICU's StringPiece.

The webkit change (which is minor) will be dealt with in the webkit bugzilla.

This can go in before the ICU change/upgrade without affecting anything. 


BUG=8198
TEST=All the targets are built without an error on all platforms.

Review URL: http://codereview.chromium.org/171012

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@23613 0039d316-1c4b-4281-b951-d872f2087c98


CrOS-Libchrome-Original-Commit: b5b2385af64eb08b1feb53fb0ad65c835f472912