Fix race in directory creation on Windows.  The file_util::CreateDirectory() function is supposed to succeed if the directory already exists, but in the previous implementation when two processes/threads called the function at the same time for the same path, one would see a failure.
Also, get rid of the use of SHCreateDirectoryEx function.  For one, it
is highly serializing because it calls SHChangeNotify.  The race was
actually easy to reproduce by starting two instances of Chrome both at
the same time, with a profile directory flag indicating a profile
directory that did not previously exist.

For another, SHCreateDirectoryEx would fail with ERROR_CANCELLED if
any of the path components were not visible, according to its MSDN
documentation.

BUG=none
TEST=[base_unittests.exe --gtest_filter=*CreateDirectory* ], and/or install Chrome Frame on a fresh machine and make sure the first page you navigate to has two Chrome Frame instances on it - you should get no DCHECK on failure of PathService::Override

Review URL: http://codereview.chromium.org/437090

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@33225 0039d316-1c4b-4281-b951-d872f2087c98


CrOS-Libchrome-Original-Commit: 40676ab2827e430e5719b1646aa66e9d07bc37fe