Don't reuse the initial IPC channel.
I was seeing a non-initial IPC channel getting closed in a renderer, and then
when someone tried to reuse that channel by name,
IPC::Channel::ChannelImpl::CreatePipe would assign the initial pipe.  The
initial pipe was already in use, and things would fall apart pretty rapidly.

I'm making this FATAL because the renderer's probably going to be unusable if
it gets into this state anyway, and a sad tab is probably more useful than a
tab that appears to be loading indefinitely.

BUG=26754
TEST=Test case from bug 26754 comment 9 (affected machines only):
       a. Have lots of bookmarks (import Safari defaults)
       b. Rightclick on bookmark bar, and choose "Open All Bookmarks"
      Expect: no crash.
Review URL: http://codereview.chromium.org/452021

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@33351 0039d316-1c4b-4281-b951-d872f2087c98


CrOS-Libchrome-Original-Commit: 554a8858ada95672d470a14668d5854f47a4ebd4