Refactor FileVersionInfo into an interface with platform implementations. This allows us to move the chrome specific version informaton used by Linux into src/chrome.
Add a GetChromeVersionInfo() for Linux in src/chrome/app/ and make sure to use this in src/chrome.

In src/webkit/glue, add a new glue method for getting the product version.  When compiling chrome, use an implementation in src/chrome/renderer (which uses GetChromeVersionInfo()) and a stub implementation for test_shell.

Review URL: http://codereview.chromium.org/1560027

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@44435 0039d316-1c4b-4281-b951-d872f2087c98


CrOS-Libchrome-Original-Commit: bcff05af3d60df10a21e07751a14b95ce9e7f3d4