Move the SingletonSocket to a temporary directory
This is to workaround problems on certain network filesystems (notably AFS)
which do not support Unix domain sockets. We move the sockets into a temporary
folder and symlink. To avoid the possibility of a dangling link to a missing
(and later intercepted) remote directory, we create and check cookie files
and rely on the stickiness of /tmp/ to avoid a race condition in the check.

R=mattm
BUG=44606
TEST=ProcessSingletonLinuxTest.*

Review URL: http://codereview.chromium.org/2838034

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@57623 0039d316-1c4b-4281-b951-d872f2087c98


CrOS-Libchrome-Original-Commit: 53f4826c9fe6bea1718a499dc686f26dbf24ab50