Allow histograms in cache stats.cc to bypass corruption checks
This is a relanding of http://codereview.chromium.org/4174002

The change in this CL is that I've made the function
to look for errors virtual, and over-rode it in StatsHistograms.
The current override doesn't actually check for inconsistencies,
so that the UMA checks (and renderer-to-histogram checks) won't
complain and refuse to transport the histograms.

BUG=61281
r=rvargas
Review URL: http://codereview.chromium.org/4364001

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@64990 0039d316-1c4b-4281-b951-d872f2087c98


CrOS-Libchrome-Original-Commit: 93a41d7ecc35fe54ed7182c1900de1e1461caea3