Changing the installer switches from wchar_t[] to char[].
Because of this I'm also refactoring some code that before
was using wstring to build command lines by hand instead of
using the CommandLine class.  Now we use CommandLine.

To get this to work correctly, I also needed to fix CommandLine::AppendArguments so I added a little test for it.

TEST=There should be no changes in functionality.  Run all installer tests.
BUG=61609

Review URL: http://codereview.chromium.org/4928002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@66088 0039d316-1c4b-4281-b951-d872f2087c98


CrOS-Libchrome-Original-Commit: 12c0b1823a10d2b0cf63144bcd32b556c73177af