Pull in ICU 4.6
Adjust the expected result for the display name of es-419. It does not have
'Carribean Islands' any more in the English display name.  Also, swap
two parameters of EXPECTED_EQ so that the 1st param is 'expected' and
the 2nd is 'actual' in two tests in l10n_util_unittest.cc

In addition, change icu_string_conversions.cc to fix the failure
of ICUStringConversionsTest.ConvertCodepageUTF8.
The test uses WideToCodepage that calls u_strFromWCS. Where
U_WCHAR_IS_UTF32 is defined, it calls u_strFromUTF32. On Mac,
it's not defined because neither __STDC_ISO_10646__ nor _UCS4_ is defined.
As a result, wcstombs is called, which doesn't work as we want it to.

There are two ways to fix this:
 - Directly call u_strFromUTF32 on Mac/Linux. What this CL does.
 - Manually define U_WCHAR_IS_UTF32 in icu's pmac.h, which is done
   in http://codereview.chromium.org/6578003

We do both to be block wcs*mbs/mbs*wcs from being called in other parts
of ICU when U_WCHAR_IS_UTF32 is not defined.

BUG=61514
TEST=Build goes fine and tests pass.
Review URL: http://codereview.chromium.org/6532030

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@75854 0039d316-1c4b-4281-b951-d872f2087c98


CrOS-Libchrome-Original-Commit: 79157356c6481b608970379e00cded811e00c003