touch: more fixes in the message pump.
Remove the signal handler when the message-pump terminates. This wouldn't
normally be necessary, but it can happen that the message-pump is terminated,
and then additional GTK+ operations happen (e.g. in unit_tests).

Also check for non-null dispatcher before dispatching events. This can happen
for the first-run dialog, and also in the unit_tests.

BUG=none
TEST=unit_tests

Review URL: http://codereview.chromium.org/6596094

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@76643 0039d316-1c4b-4281-b951-d872f2087c98


CrOS-Libchrome-Original-Commit: 4b9baf9d5a55852ab9613a04d745e891a6d083f0