Revert 76239 - Crash when we notice a corruption of the histogram range-vector
Reason for revert: This patch might have resulted in some
flakiness in the sync bots on the chromium waterfall.

The range vector is calculated once when a histogram is 
constructed, and should never change over the lifetime of the
histogram.  When it does change, it means that memory is being
overwritten.  We now crash (via CHECK) when there is any 
detected corruption of the range vector.

This CL uses a more robust check-sum algorithm to detect corruption 
of the vector.  The previous algorithm just did a sum, and was
too easilly tricked by small (offsetting) changes in several
ranges.  Hopefully, this CRC-32 implementation will not be
so easilly fooled.  

I had to refactor the class to ensure that each histogram was completely
constructed before I registered it.  The old code could sometimes
register a base class (tradtional Histogram, with exponential
bucket spread) and then run the derived constructor (such as
when creating a LinearHistogram, with a second construction 
of the ranges and checksum).  I now carefully avoid
generating the checksum until fully constructing the instance,
 and then I run InitializeBuckets only once on
the instance before registering it.

bug=73939,74467
r=mbelshe
Review URL: http://codereview.chromium.org/6577013

TBR=jar@chromium.org
Review URL: http://codereview.chromium.org/6611001

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@76667 0039d316-1c4b-4281-b951-d872f2087c98


CrOS-Libchrome-Original-Commit: 9bf0f906c4c5fd954ed424f52bc2d2ee48cf538c