Reland code to detect corruption of histogram ranges
This code was reverted because it caused, or exposed
flakiness in sync unit test code (mostly on the Mac).
We need to monitor that closely as we re-land.


See the reviews and history in CLs:

http://codereview.chromium.org/6577013
and
http://codereview.chromium.org/6591052

This code generates CRC checksums for the histogram
bucket-range vectors, and guarantees they remain
intact when the histogram data is about to be
uploaded to UMA.  If the data is corrupted, then
we will fail on a CHECK().


r=mbelshe,rvargas
bug=73939

Review URL: http://codereview.chromium.org/6627011

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@77033 0039d316-1c4b-4281-b951-d872f2087c98


CrOS-Libchrome-Original-Commit: 98359ce4f10e42fd1ff8cb50dbf8d26a1ce01fa5