Remove the sum of square tally from histograms
This value is accrued without locking, and hence is too
commonly corrupted.  Even if it is gathered
perfectly, I've been informed that for very large
samples the computed variance is "not a stable
metric" and can diverge."  If we did want to 
look at the variance for most distribution, then
we need to look at the variance of the log
normal distribution, which would mean we wanted
to take the sum-of-squares of the log, which is
computationally intense... and probably of little
value.

Bottom line: Remove the useless metric.

I still need to change the protobufs so that
this value is optional in uploads, but other than
that tiny point, this CL removes almost all
traces of this metric.


r=rtenneti
Review URL: http://codereview.chromium.org/6672071

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@79225 0039d316-1c4b-4281-b951-d872f2087c98


CrOS-Libchrome-Original-Commit: 0704a97d5caa9b7d90d2d9ed39b3db91557a3cb0