Move flags initialization into block before sharing across threads
We re-initialized the flags after we shared the histograms, so
there was a potential race for multiple simultaneous initializations.
This change completes the basic setting of the flags prior to
registration, which in turn shares the histogram across threads.

There was a potential race with the (later) setting of the flag
during an upload from the renderer.  With the new code, the flag
is const except for its potential change by the renderer-serialization
routine, which is also the only consumer of this flag (in
the renderer).

I also made the corresponding change to stats_histogram, even
though the code was already at least as correct, as I wanted
the patterns of use to remain analagous.

r=rvargas,rtenneti
BUG=77760
Review URL: http://codereview.chromium.org/6731074

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@79835 0039d316-1c4b-4281-b951-d872f2087c98


CrOS-Libchrome-Original-Commit: 05a81412f30cc6ee91dc2ffb3205f24deaaac54d