Revert 83641 - DNS Host resolver changes with retry logic. Fix for bug Chromium cannot recover from a state when its DNS requests have been dropped.
Whenever we try to resolve the host, we post a delayed task
to check if host resolution (OnLookupComplete) is completed 
or not. If the original ateempt hasn't completed, then we start
another attempt to resolve for the same request.

We take the results from the attempt that finishes first and
leave all other attempts as orphaned.

BUG=73327
TEST=dns host resolver tests
R=eroman,jar
Review URL: http://codereview.chromium.org/6782001

TBR=rtenneti@chromium.org
Review URL: http://codereview.chromium.org/6902198

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@83710 0039d316-1c4b-4281-b951-d872f2087c98


CrOS-Libchrome-Original-Commit: e57ec4c1653c7f08cf6d92f3341a04f242885814