[Mac] Implement the system dictionary popup by implementing NSTextInput methods.
This is a two-sided patch; the Chromium side is plumbing to marshall data from
the renderer to the system APIs.

Note that just hitting Cmd+Ctrl+D usually does not bring up the popup. I think
this may be an Apple bug, but I have not yet found a work-around.

BUG=17951,37715,47141
TEST=Hold Cmd+Ctrl+D on a web page and mouse around. The dictionary popup should follow the mouse and show the definition of the current word.
TEST=In a text area, the dictionary popup should work only if the text area has focus.
R=avi,suzhe,jam

Review URL: http://codereview.chromium.org/6289009

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@83723 0039d316-1c4b-4281-b951-d872f2087c98


CrOS-Libchrome-Original-Commit: d4cff27a73a9dc6caa6b37a8b28ef72ecc35fa0d