IPC: Increase MAX_DESCRIPTORS_PER_MESSAGE
This fixes the loading of Native Client's integrated runtime (IRT)
library on Mac, because we need to send 5 descriptors in NaCl's
startup IPC message on Mac.

I am being conservative in only increasing MAX_DESCRIPTORS_PER_MESSAGE
by 1 because it carries a storage cost (in
Channel::ChannelImpl::input_cmsg_buf_) which is multiplied by the
number of messages that can fit into Channel::kReadBufferSize, and the
storage is allocated for the lifetime of the Channel.

Note that I am not increasing the size of input_cmsg_buf_ under "#if
defined(OS_MACOSX)" in ipc_channel_posix.h because it is already too
small in the general case (see http://crbug.com/82829), but NaCl does
not exercise this general case.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=1804
TEST=NaCl's "run_ppapi_core_browser_test irt=1" when run against Chromium

Review URL: http://codereview.chromium.org/7010015

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@85617 0039d316-1c4b-4281-b951-d872f2087c98


CrOS-Libchrome-Original-Commit: 49d37ef9a5f31e0c7bc61750215957e702572c78