Add Bus::ShutdownOnDBusThreadAndBlock() and remove bus::Shutdown()
This function is intended to use at the the very end of the browser shutdown,
where it it makes more sense to shut down the bus synchronously, than trying
to make it asynchronous.

Remove Bus::Shutdown() as we are unlikely to need it for the production code.

BUG=chromium:90036
TEST=dbus_unittests


Review URL: http://codereview.chromium.org/7830009

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@99331 0039d316-1c4b-4281-b951-d872f2087c98


CrOS-Libchrome-Original-Commit: e20d39fe8164f6633d8224030133a1c6ad18a289