Use NeedsScopedRefptrButGetsRawPtr instead of UnsafeBindtoRefCountedArg.
Will's implementation of NeedsScopedRefptrButGetsRawPtr checks for refcounting by testing for converability of the pointer values to base::RefCounted and base::RefCountedThreadsafe.  This happens to tolerate incomplete types in a sane way on linux where the check is enabled.

UnsafeBindtoRefCountedArg's impelmentation relies on introspecting the type for the existence of a member function.  Though this will catch refcounted objects with a refcount API that are not derived from base::RefCounted and base::RefCountedThreadsafe it requires complete types which places too strong of a constraint on what kinds of arguments can be bound.

Since the majority (all?) refcounted types in Chromium are derived from base::RefCounted and base::RefCountedThreadsafe, this should be just as good of a guarantee.

Bug filed to see if there's a way to adapt UnsafeBindtoRefCountedArg.

BUG=94995,98732
TEST=new unittest.


Review URL: http://codereview.chromium.org/8101008

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@103627 0039d316-1c4b-4281-b951-d872f2087c98


CrOS-Libchrome-Original-Commit: 8217d454b2f96847028e9b0dbe7d82d66959ecb4