Create scanning script to determine header dependencies.
This python script will take a set of input files and header include directories and scan through the sources trying to find all header references to generate a list of file dependencies.  The scanner does not use a preprocessor so all headers in the source even if conditionally compiled out, are returned.  Files which can not be found in the file system are ignored.


BUG= http://code.google.com/p/chromium/issues/detail?id=96782
TEST= manual
Review URL: http://codereview.chromium.org/8037013

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@103675 0039d316-1c4b-4281-b951-d872f2087c98


CrOS-Libchrome-Original-Commit: 2cec759bc34b3f3e7fb3a624474f8ea188a3c975