dbus: slight cleanup to Property code
Correct comments based on practice that worked best for the bluetooth
adapter and device clients, in particular adding the virtual destructor
to the struct and cleaning up the return values of GetObject() in the
documentation.

Drop VLOG() calls, there's no need to debug at this level, you can use
dbus-monitor to see the incoming and outgoing calls to Chromium and can
add logging in upper layers when there are real values to log.

BUG=none
TEST=built on try bots

Change-Id: I2ac6b55e288626b5f9cd843935887e261d386dec


Review URL: http://codereview.chromium.org/9418035

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@122522 0039d316-1c4b-4281-b951-d872f2087c98


CrOS-Libchrome-Original-Commit: 992079e0d0dd2d593bcadac28be4e2d15df75644