dbus: verify object path of incoming signals
The existing behavior, while convenient for debugging, is wrong.

D-Bus will not call any further filter functions once one returns
DBUS_HANDLER_RESULT_HANDLED, in order for the next to be called a
filter must return DBUS_HANDLER_RESULT_NOT_YET_HANDLED if it does
not handle the incoming signal.

We also can't defer this to the signal function since we have to
post that to a different thread, and return values get hard.

Since object proxies are constructed per-path, and match common
interfaces and members, this means signals must be matched on an
object otherwise only the first registered object proxy for any
client will be called, and will be called for all signals.

BUG=chromium-os:27113
TEST=ran unit tests, and manually verified existing code that uses ConnectToSignal

Change-Id: Ia4cbc064dff0421a37fe4c4b7c719acf25eb630c


Review URL: http://codereview.chromium.org/9508005

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@124357 0039d316-1c4b-4281-b951-d872f2087c98


CrOS-Libchrome-Original-Commit: 4d97f2337eacef834e523a6a8462d13ad4fbd550