Upstream Android proxy config service.
This change adds in an Android-specific system proxy provider called
ProxyConfigServiceAndroid. This is created as required from C++. In
turn ProxyConfigServiceAndroid creates a java object which provides
the link to Android's proxy changed broadcast, as well as the
functionality to read the Java system properties where the current
proxy settings are located.


Review URL: https://chromiumcodereview.appspot.com/10206014

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@140748 0039d316-1c4b-4281-b951-d872f2087c98


CrOS-Libchrome-Original-Commit: 45ab7abc0b8ae4d3a78637c158d98cd06b0a8f7a