Fix a bug in file_util::Delete() where symlinks are not handled right
With this patch, file_util::Delete() uses lstat64() instead of stat64(), so that Delete() checks the existence of the symbolic link itself not a file pointed by the symlink. This allows us to delete a symbolic link when it points to a non-existent file. 

BUG=119430
TEST=base_unittests::FileUtilTest.* passes.

Review URL: https://chromiumcodereview.appspot.com/10690047

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@145060 0039d316-1c4b-4281-b951-d872f2087c98


CrOS-Libchrome-Original-Commit: 7122db29d3965050505f7294f073d83f80ea7835