Android: remove ant environment dependencies.
The logic to calculate where the SDK should be is already done by
envsetup and made available to gyp as a variable, so just have gyp pass
this as a command line argument to ant. The behaviour is identical.

This change is to move toward making it possible to build the Android
port without having envsetup.sh sourced at build time (only at
gyp-generation time), making it more like other platforms. It also
simplifies the ant buildfiles.


Review URL: https://chromiumcodereview.appspot.com/10821105

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@149400 0039d316-1c4b-4281-b951-d872f2087c98


CrOS-Libchrome-Original-Commit: 75b558c7c2b7927191c1326701f0c4f09ade385b