[Win] Don't call GetLongPathName() in CreateTemporaryFileInDir().
GetLongPathName() can fail if the user doesn't have the necessary
privileges on an ancestor directory. If this call failed during
CreateTemporaryFileInDir(), the temporary file that was created
in the previous GetTempFileName() call will leak.

None of the callers rely on the path being expanded on return. So
get rid of the call.

BUG=155612

Review URL: https://chromiumcodereview.appspot.com/11300011

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@167189 0039d316-1c4b-4281-b951-d872f2087c98


CrOS-Libchrome-Original-Commit: e1af0d5a7396136049d8a684cc88fb1d7f10aec8