android: Update comment about invalid timestamps from SurfaceFlinger
A frame with a timestamp of INT64_MAX actually means that the fence
object associated with the frame is still pending or that the frame has
not yet been displayed. Update the comment in SurfaceStatsCollector to
make this more obvious.

No functional changes.

NOTRY=true


Review URL: https://chromiumcodereview.appspot.com/13188002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@191158 0039d316-1c4b-4281-b951-d872f2087c98


CrOS-Libchrome-Original-Commit: 6afbb5dd36891ec485d02c224126bc05c025ea26