Delete CopyRecursiveDirNoCache from test_file_util.
This function was used in only one place and that place was wrong.

The implementation was long and complicated on Windows, and long and
complicated and copied verbatim from elsewhere with 3 lines different on Posix.

The place it was used was in the proxy launcher when copying the profile. It's
not clear to me why we wouldn't want the profile files in the filesystem cache
when running tests. Quite the opposite, we want the tests to run as fast as
possible.

The only place this should matter is in the startup tests. And the startup
tests do things to the profile after it gets copied that should page some files
back in! This adds another step to the startup tests to evict the profile files
for cold startup tests only.

BUG=

Review URL: https://codereview.chromium.org/13394003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@191854 0039d316-1c4b-4281-b951-d872f2087c98


CrOS-Libchrome-Original-Commit: a6d054e0fbebac6ef52c2d1587c2acced87c10f7