D-Bus: allow multiple signal handlers for a signal
For the org.freedesktop.DBus.Properties.PropertyChanged signal, all
relevant clients will share a single ObjectProxy since they share
the same path and interface; the actual destination client for the
signal is determined by its arguments.

This means that we must support multiple signal handlers for a single
object proxy, the previous fix of replacing with the latest was not
sufficient for this case.

Due to issue 223483, this is not a complete fix. Property objects
coming and going will leave signal handlers in the ObjectProxy with
NULLd weak pointer references - they will be harmless, but use up
memory.

BUG=234380
TEST=dbus_unittests
R=satorux@chromium.org

Review URL: https://codereview.chromium.org/14333009

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@195953 0039d316-1c4b-4281-b951-d872f2087c98


CrOS-Libchrome-Original-Commit: 012e781dc8f619507b01a5d6f7517d5c9cca39b8