Delete CopyRecursiveDirNoCache from test_file_util.
This function was used in only one place and the implementation was
copy-and-pasted. This replaces the call with a cross-platform implementation in
the one file that needs it.

This is a reland of r194664 which was a reland of r192940. This version does
not attempt to fix the startup test caching bugs I discovered when rewriting
the function. I filed bug 237904 for this.

BUG=

Review URL: https://codereview.chromium.org/14577009

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@198539 0039d316-1c4b-4281-b951-d872f2087c98


CrOS-Libchrome-Original-Commit: 6a48fb51b8f5f627781177ec8af1cf1c56a03abe