Define EncodedVideoSource and RtcCapturedEncodingVideoCapturer.
Defined the interfaces for EVS (EncodedVideoSource) and related IPC messages
between renderer and browser processes for the purpose of encoded screen capture.

Added RtcCapturedEncodingVideoCapturerFactory which implements the
libjingle::WebRtcVideoEncoderFactory class, and the corresponding
RtcCapturedEncodingVideoCapturer which implements the webrtc::VideoEncoder.

Modify MediaStreamDependencyFactory and inject RtcCapturedEncodingVideoCaptureFactory
in PeerConnectionFactory. For the time being, the factory will only be instantiated
for Chrome OS and ARM.

BUG=238515
TEST=trybot

Review URL: https://chromiumcodereview.appspot.com/16320005

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@209760 0039d316-1c4b-4281-b951-d872f2087c98


CrOS-Libchrome-Original-Commit: 76bf7f58d93256dfa5867b6730e446517ead2a30