Add a HistogramRecorder class and use cases.
This CL adds a utility class to streamline the task of testing that metrics have been updated as expected. Specifically, the HistogramRecorder class allows a client to obtain the differential value of a given histogram in the interval since the given HistogramRecorder instance was created.

BUG=232414

Review URL: https://chromiumcodereview.appspot.com/18337014

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@212378 0039d316-1c4b-4281-b951-d872f2087c98


CrOS-Libchrome-Original-Commit: 3a0e3d498cdd0be4659e03dd05888b4ae15f4993