Pass app URL to mojo_shell on Android
This CL lets us supply an app URL to mojo_shell on Android. To load an app, you
can use the adb_run_mojo_shell script:

$ ./build/android/adb_run_mojo_shell http://127.0.0.1:4444/libsample_app.so

This script starts the MojoShellActivity on the device and passes the URL to
mojo_shell via an Android Intent. We're currently using the VIEW intent, but we
might want to switch to a different intent at some point.

R=darin@chromium.org,aa@chromium.org

Review URL: https://codereview.chromium.org/47343002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@231260 0039d316-1c4b-4281-b951-d872f2087c98


CrOS-Libchrome-Original-Commit: e5fc9bfffb15f8d85590e5e123e65260774f31e3