Initial cut at object to listen for handle to be ready
It creates a singleton background thread that uses MojoWaitMany() to
wait for a handle to be ready. An additional communication pipe is
used to wake up the thread as necessary (when set of handles changes,
or to quit).

Currently it only waits for ready to be read from. If we need the
ability to listen for writes I can add that.

BUG=none
TEST=none
R=darin@chromium.org

Review URL: https://codereview.chromium.org/56333002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@232716 0039d316-1c4b-4281-b951-d872f2087c98


CrOS-Libchrome-Original-Commit: 03be76c122dee5b086208950c216824b80e5f37f