Android: remove ContentViewCore from the renderer crash codepath.
This patch employs base::TerminationStatus to communicate the state of the oom
bindings when a child process dies and exposes renderProcessGone() in
WebContentsObserverAndroid, removing the need for ContentViewCore to be
involved in the renderer crash codepath.

This allows to drop tracking the pid of the renderer process in ContentViewCore
and eliminates the need for the |tab_crashed_| flag in native
ContentViewCoreImpl.

BUG=314583

Review URL: https://codereview.chromium.org/58203002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@233905 0039d316-1c4b-4281-b951-d872f2087c98


CrOS-Libchrome-Original-Commit: 6bf2f78736b7165e2762806f001c36425b9f0e6a