Android: relands r237947, split golden content from jni_generator_tests.
Rather than intermix the generated golden files with the test script,
extract in separate files.
Also, if "REBASELINE" environment variable is set, it'll re-generate the
golden scripts, to it's much simpler to do changes in the generator itself.
Note: there are NO functional changes in the generator itself, only in the tests.

Previous attempt failed with mismatch copyright message due to the precommit
hook.

BUG=

Review URL: https://codereview.chromium.org/98573004

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@238084 0039d316-1c4b-4281-b951-d872f2087c98


CrOS-Libchrome-Original-Commit: 3620b73d2649e0974e810c05dc3d41d45ebe098f