Disable timing from chrome://profiler on Android.
On Android, on a arm CPU, TimeTicks::Now() is taking 900ns (20 times
what it takes on a desktop x86 CPU). The bigger user of TimeTicks::Now()
is tracked_objects::TrackedTime, and this impacts the overall
performance of Chrome on this platform.

This CL adds compile line flags that allows to disable
tracked_objects::TrackedTime. It also disable
tracked_objects::TrackedTime by default on Android.

BUG=315070

Review URL: https://codereview.chromium.org/99343002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@240092 0039d316-1c4b-4281-b951-d872f2087c98


CrOS-Libchrome-Original-Commit: 915b344f5c907df65e6f21cda93d6f7859b1f853