compiler_specific.h: define FINAL as final when using Clang on Windows
Clang on Windows sneakily identifies itself as both Clang and MSVC :)
Previously, we would end up defining FINAL as "sealed" here, but in Blink,
we'd define it as "final", causing lots of warnings. This change means
we define it the same in both places.

BUG=82385
NOTRY=true

Review URL: https://codereview.chromium.org/134243005

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@248499 0039d316-1c4b-4281-b951-d872f2087c98


CrOS-Libchrome-Original-Commit: 38950138d995f881cf9140ba73900128a93122c1