Implement gpu::CommandBuffer over mojo IPCs and have mojo::gles2 use that.
This implementation is not complete yet, but is enough to get all the samples
working.
There is one major hack which is how to pass shared memory handles across mojo
IPC. We will need to add the right capability and/or rework the command buffer
protocol to work on top of mojo primitives.

BUG=333157
R=darin@chromium.org, jamesr@chromium.org

Review URL: https://codereview.chromium.org/149333004

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@249556 0039d316-1c4b-4281-b951-d872f2087c98


CrOS-Libchrome-Original-Commit: 68780860e0e5ff26d9d2a8af6dc97539985062f9