[Android] Disable tombstones for renderer crashes on JB MR2+
Starting with JB MR2, a crashing service triggers a dialog. This is an
awful user experience when a Chrome renderer crashes. The system service
that triggers the dialog also prints the stacktrace to the log and
creates a tombstone.

For now, disable the tombstone and dialog on JB MR2+ "user" builds (or
rather, non-"userdebug", non-"eng" builds).

BUG=273706
TEST=navigate to "about:crash". If on JB MR2+ "user" build, no system
dialog should be shown. Otherwise, dialog should be shown.

Review URL: https://codereview.chromium.org/113873006

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@251183 0039d316-1c4b-4281-b951-d872f2087c98


CrOS-Libchrome-Original-Commit: a588c095b4b3b99941eb735d46e8153362bb82d5