Mojo: add scaffolding for request/response support
Introduces MessageHeaderWithRequestID, which extends MessageHeader, and a corresponding MessageWithRequestIDBuilder.

Adds an AcceptWithResponder method to MessageReceiver. The idea is that this will be used by generated proxy code to kick off a request that expects a response. Somehow we will bind a user supplied callback to a MessageReceiver::Accept implementation. Details TBD.

Internally, Connector might get broken down into two layers: a core pump layer (the code that currently exists) and a router layer that'll be responsible for routing incoming messages to the proper MessageReceiver, either a registered responder or the default incoming receiver.

R=davemoore@chromium.org

Review URL: https://codereview.chromium.org/185803002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@254541 0039d316-1c4b-4281-b951-d872f2087c98


CrOS-Libchrome-Original-Commit: 3714c8b156378fb4ce5357ea28e73c7e2f255a63