Add ScopedGeneric.
This is intended to be used in a typedef for declaring various types of
scoped resources, like ScopedFD or ScopedGDIObject. It attempts to be like
scoped_ptr but without pointer semantics.

Currently the scoped objects are either custom one-offs that don't support move
semantics and have subtly varying behavior, or they use a pointer like ScopedFD
which is a memory stomp waiting to happen (since you must keep the int* alive
longer than the scoper so it can be dereferenced and closed).

R=ajwong@chromium.org, viettrungluu@chromium.org, ajwong

Review URL: https://codereview.chromium.org/189613002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@256596 0039d316-1c4b-4281-b951-d872f2087c98


CrOS-Libchrome-Original-Commit: 821261bc506fe4a04a0ae0ce8eb6488cca74e52d