Convert codec.js to use a DataView instead of a Uint8Array.
codec.js is modified to use a DataView as its primary means of accessing
an arraybuffer as opposed to a Uint8Array. The DataView then handles the
signed/unsigned case appropriately for us without any bit manipulations.

We can then remove the load*/store* functions, execpt for the 64-bit
versions which remain to pack to/from JS doubles.

This then introduces a need to know the endianness of the host, since Mojo
messages are written in native byte oreder, but DataView wants a specific
ordering (with a big-endian default).

The IDL code generator for JS is modified to call into the new functions,
which are renamed to avoid confusion between the signed and unsigned versions.

The WebUIMojoTest.EndToEndEcho is modified to pass values for each of
the possible integral types, to check for correct signedness on the
javascript side, and to run fewer iterations.

BUG=

Review URL: https://codereview.chromium.org/242613010

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@265473 0039d316-1c4b-4281-b951-d872f2087c98


CrOS-Libchrome-Original-Commit: 441c47f2715a31ea39ab95806a70f8067395ed08