Make android bot configuration more consistent.
On the chromium.perf waterfall, we have some devices that are not going offline
(N4 + N5) and some that are (N7, N10, GN). Diffing the results of the
provision_device step pointed out some configuration differences between the
working and non-working devices. Most notably, wifi_on was set on all the
devices that are going offline and not on the ones that are staying on.

I hope making these settings deterministic will get these devices into better
shape.

BUG=366697

Review URL: https://codereview.chromium.org/251793003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@266394 0039d316-1c4b-4281-b951-d872f2087c98


CrOS-Libchrome-Original-Commit: 58de77d09564a43ae1934e2cee4543bc028c25c6