Change the way libc++ is linked when use_custom_libcxx=1.
Instead of passing libc++.so and libc++abi.so on the command line to the linker,
we want to be able to simply pass -stdlib=libc++ to the clang driver and let it
make the decision about linking the DSOs. In this way we can avoid linking
libc++ into C executables.

To this effect we introduce a proxy target which adds -stdlib=libc++ to direct
dependents' flags but does not link the actual DSO. We also link libc++abi
statically into libc++. The latter change is necessary because clang's
-stdlib=libc++ mode is set up to use libsupc++ as the ABI implementation, and
will not automatically pick up libc++abi.so.

BUG=318770
R=glider@chromium.org, thakis@chromium.org
NOTRY=true
TEST=msan

Review URL: https://codereview.chromium.org/264613005

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@268169 0039d316-1c4b-4281-b951-d872f2087c98


CrOS-Libchrome-Original-Commit: 121cbf6ec94489a2bd430ca2f54a5d7d520df113