Crazy linker fix for
https://code.google.com/p/chromium/issues/detail?id=355595

There is a race between the crazy linker running on the native thread and the system linker running at the same time on the main UI thread,  It can show up in several places, but in its simplest incarnation appears as:

 --- crazy linker on native thread
 mprotect page writable
 --- system linker on main thread
 mprotect page writable
 write it
 mprotect page readonly
 --- crazy linker on native thread
 write it                [ <- sigsegv, page readonly (code=2) ]
 mprotect page readonly  [ not reached ]

The patch moves the crazy linker's r_map update calls so that they execute on the main UI thread.  This ensures that the crazy linker sets these pages writable and then readonly once done synchronously with system linker activity.

Companion to:

https://googleplex-android-review.git.corp.google.com/#/c/448556/

BUG=355595

Review URL: https://codereview.chromium.org/228023005

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@270084 0039d316-1c4b-4281-b951-d872f2087c98


CrOS-Libchrome-Original-Commit: 284ae62aa0ed0f836876ce4cc2a3824b7896bf0b