Extend PR_ParseTimeString() to accept some ISO 8601 formats to fix timezone parsing in SyslogParser.
Previously, SyslogParser::ParseTime() would parse time, reformat it in a way that is understood by NSPR and then hand it off to PR_ParseTimeString() where it is parsed again.  This double-parsing is unnecessarily convoluted and slow.  Extending PR_ParseTimeString() simplifies and speeds up SyslogParser::ParseTime() which is time-critical in case of large log files.

BUG=370509

Review URL: https://codereview.chromium.org/266193002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@270815 0039d316-1c4b-4281-b951-d872f2087c98


CrOS-Libchrome-Original-Commit: c3701ebc95bef612f6cb8250390f348aae4cd0a1