Introduce internal::ServiceRegistry to prepare for ServiceProvider split.
This is to get ready to split ServiceProvider into Shell and Application. It makes the existing Application no longer subclass internal::ServiceConnectorBase::Owner. Instead it puts that functionality in the new class internal::ServiceRegistry and Application uses that by composition.

The validation in this cl is meant to be consistent with what's currently in the tree. I'll redo the cl that adds the ability to add a service with a specific validator after the split is done.

BUG=

Review URL: https://codereview.chromium.org/337533002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@276859 0039d316-1c4b-4281-b951-d872f2087c98


CrOS-Libchrome-Original-Commit: 183a0500d63464a642fbdfd459b3a7f0a7873482