Expand the JsonPrefStore's interface to allow for an alternate filename to be specified.
The alternate filename is intended to be the old name of the pref file being read, JsonPrefStore handles the migration if required. JsonPrefStore's FileThreadDeserializer is the best place to do this as it's the only code in the prefs stack running on a thread allowed to do IO (another solution would be to let JsonPrefStore take a PreReadOnFileThreadCallback to abstract the required work away, but the direct approach seemed easier here without breaking encapsulation).

BUG=372547

Review URL: https://codereview.chromium.org/347793002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@278776 0039d316-1c4b-4281-b951-d872f2087c98


CrOS-Libchrome-Original-Commit: cfcf0e53fa4ee2f4b89b477390f1906db0cbf4d9