Reland change to use a native MessagePump instead of a MessagePumpDefault. Original CL: https://codereview.chromium.org/331513002/
The previous CL caused a bunch of LibJingle unit tests to fail because the tests
were using a gtest death test which runs in the context of a fork()ed but not
execu()ed process.

A CF Message loop can't live in this environment.  Explicitly construct a
MessagePumpDefault for these tests.

BUG=356804,385604

Review URL: https://codereview.chromium.org/331983002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@279013 0039d316-1c4b-4281-b951-d872f2087c98


CrOS-Libchrome-Original-Commit: 3669d7d8f64445139731e73da773dbd2d4a81afb