Allow the forwarder to recover if the device_forwarder crashes or hangs
I've noticed on the perf bots that if we get a host_forwarder error
because it can't connect to the device, then all subsequent forwarder
operations will fail. Example:
http://build.chromium.org/p/chromium.perf/builders/Android%20Nexus4%20Perf/builds/1163/steps/Sharded%20Perf%20Tests/logs/stdio

This patch fixes that by killing the device forwarder (and ensuring it
isn't initialized) so that it can be re-initialized on the next
attempt.

BUG=294878

Review URL: https://codereview.chromium.org/427353005

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@287105 0039d316-1c4b-4281-b951-d872f2087c98


CrOS-Libchrome-Original-Commit: 5d249080252aaa28b615af92132a8423fc857d46