Fix a DCHECK violation in ipc_perftests
We only allow one MessageLoop instance alive at any given time on a thread. If
we use IPCTestBase::set_message_loop(), we need to do it in two steps: first set
the message loop to NULL, and then set it to a newly-created message loop.

Therefore, I changed the interface of IPCTestBase to make it a little
more friendly.

BUG=None

TEST=ipc_perftests should not crash in debug build.

Review URL: https://codereview.chromium.org/538593002

Cr-Commit-Position: refs/heads/master@{#293439}


CrOS-Libchrome-Original-Commit: 22796d8d5fd362111b7aa431d229aa2ffc91aedd