Migrate ResourceExtractor.java inside org.chromium.base package.
With the CL, https://codereview.chromium.org/456413002, we are migrating from embedding inline resources
inside blink towards using chromium's grd resource system. As part of this effort, most of the tests
would require to use the 'pak' files packaged as android application assets.

These assets need to be extracted to the application data directory during the startup of the test
application. The native libraries would be able to work with these 'pak' files only after the extraction
from the asset manager.

To achieve this goal, we would require that ResourceExtractor.java to be moved inside the org.chromium.base
pacakge so that it can be utilized in ChromeNativeTestActivity.java

BUG=312586

Review URL: https://codereview.chromium.org/548023002

Cr-Commit-Position: refs/heads/master@{#294548}


CrOS-Libchrome-Original-Commit: e9bb6910f700e0943c157f6802b1ae3793bab221