Creating infrastructure for profiler instrumentation only in developer build and Canary channel.
This introduces TrackingProfile class that is equivalent to ScopedProfile what tracking is enabled and does nothing otherwise.

There is an idea to remove the single-parametered ScopedProfile constructor, but for this CL, I’m leaving it. Otherwise, I’d have to replace ~100 ScopedProfile instantiations throughout the code with TrackingProfile, and the substantial changes would be lost in these changes. I’ll do this in a separate CL. I replaced one ScopedProfile just to make sure this will compile.

BUG=401560
TBR=asanka@chromium.org

Review URL: https://codereview.chromium.org/663093005

Cr-Commit-Position: refs/heads/master@{#301695}


CrOS-Libchrome-Original-Commit: 20175533a430be7798d031992b83a0b4f043e939