Enable ChannelMojo on Linux
This change enables ChannelMojo on Linux.
There could possibly be a perf slowdown, especially
on latency metrics.

If you're looking for the cause of some perf bot redness,
probably this is that one. Please feel free to revert and
let me know which benchmark was hit.

BUG=664413003
R=viettrungluu@chromium.org

Review URL: https://codereview.chromium.org/684323002

Cr-Commit-Position: refs/heads/master@{#301992}


CrOS-Libchrome-Original-Commit: d4d85c640375ea7f9692e53860f4de058a7f0ea1