Make landmines handle GN build directories better.
GN has user-managed build output directories, and the build arguments are saved in those directories. This means that when you hit a landmine, the user will be forced to regenerate that build directory, and to respecify their build arguments.

This change makes the landmine script look for GN build directories. If one is found, the build and command-line arguments to GN are preserved, and enough of a bootstrap build is left in that directory to regenerate itself the next time the build is run. This should make landmines transparent to GN users (except of course that everything will be rebuilt).

Review URL: https://codereview.chromium.org/709813003

Cr-Commit-Position: refs/heads/master@{#303528}


CrOS-Libchrome-Original-Commit: f5b2cda7e05f7ffa4ac57b6a00d9cd52c1e0ac87