Instrumenting callbacks to InFlightBackendIO.
Earlier instrumentations showed that the jank for this bug comes from one of completion callbacks.

This code wraps passed callbacks with instrumentations, so that we know which operation's callback causes jank.

This CL reimplements callback wrapping, which was originally implemented in a wrong way.

BUG=422516

Review URL: https://codereview.chromium.org/745853002

Cr-Commit-Position: refs/heads/master@{#305281}


CrOS-Libchrome-Original-Commit: 6452bca69783b5041329a88a012762ebc10f7a9c