Deleting the forwarding header files base/debug/trace_event*
Now that the migration of the tracing files is completed and all the
tracing clients have been updated to point to base/trace_event/,
it is safe to remove the forwarding headers from the old location.
The only purpose of those headers, in fact, was to handle the renames
in smaller CLs.
The only thing left, at this point, is gradually renaming the
namespaces, which is going to be handled in the upcoming CLs.

See crrev.com/837303004 and the related bug for motivations.

BUG=451032
TBR=jam@chromium.org,sky@chromium.org

Review URL: https://codereview.chromium.org/897663002

Cr-Commit-Position: refs/heads/master@{#314357}


CrOS-Libchrome-Original-Commit: 74a89f06ec7db49e1262847d4aba07ed87fd5930