Change ipc_perftests large-message test count to 1,000
The ipc_perftests and mojo_message_pipe_perftests results appeared
designed to be compared against each other, and the ipc_perftests
appear to be designed to take about a second each. Therefore it is
presumably a bug that the final ipc_perftests test executes only 100
times (versus 1,000 times for mojo_message_pipe_perftests) and that it
executes in ~50 ms.

This change increases the iteration count from 100 to 1,000.

Review URL: https://codereview.chromium.org/903513002

Cr-Commit-Position: refs/heads/master@{#314698}


CrOS-Libchrome-Original-Commit: 903fdfd764c0a298862a47630fa5db146bc874cb