The Certificate Web UI incorrectly formats the issued and expires dates.
The Web UI incorrectly states that the issued and expires dates for
certificates is in UTC, when in fact it's in local time. Additionally,
it appends UTC in a non-internationalized way. This patch fixes both
issues by adding a new time_formatting.h utility function to produce a
formatted time string with the local time zone appended.

BUG=426909

Review URL: https://codereview.chromium.org/827123004

Cr-Commit-Position: refs/heads/master@{#315240}


CrOS-Libchrome-Original-Commit: 62c6b3e4d72f287ff007895052e5166c6ccd79e0