Re-sort GYP files with the latest version of tools/source_list.py
In crrev.com/316164, a bug was fixed to handle comments in
a source list correctly.

In this patch, a bug was fixed to handle comments at the
end of a source list.

Along the way, sort some GYP files in sub directories under
base and net, which I forgot to sort previously.

BUG=456014
TEST=everything should build as before

Review URL: https://codereview.chromium.org/924743002

Cr-Commit-Position: refs/heads/master@{#316418}


CrOS-Libchrome-Original-Commit: eef164ef753ccb97ae45d6c86e12a3f2c3318d1d