base: Remove use of DiscardableMemoryManager from shmem implementation of discardable memory.
DiscardableMemoryManager does not make sense for browser-wide
discardable memory where memory is already managed properly
on the browser side. Using DiscardableMemoryManager in the
case of browser-wide discardable memory just makes it harder
to ensure that release the address space in a timely manner.

This removes the use of base::DiscardableMemoryManager
from shmem implementation of discardable memory and instead
uses content::DiscardableSharedMemoryManager implementation
to make sure the address space associated with a discardable
memory segment that have been purged by the browser is released
in a timely manner.

BUG=429415
TEST=content_browsertests --gtest_filter=ChildDiscardableSharedMemoryManagerBrowserTest.*, content_unittests --gtest_filter=DiscardableSharedMemoryHeapTest.ReleaseFreeMemory

Review URL: https://codereview.chromium.org/872253002

Cr-Commit-Position: refs/heads/master@{#316753}


CrOS-Libchrome-Original-Commit: 549bd27a10ae6616109d3edbcc8f481e9e04d43b