android: Refactor function name template in JNI generator.
Factor out the code which generates the native export stub name for a
given native function, previously duplicated in two places, and also use
it in a third place: when generating the table of method registrations.
This will allow the table to be generated even when in native export
mode, as previously the table referred to the default name which is not
defined when using native exports.

BUG=442327

Review URL: https://codereview.chromium.org/938373002

Cr-Commit-Position: refs/heads/master@{#317302}


CrOS-Libchrome-Original-Commit: 37d21163c4d4d4079747a681b069d124b373d58f