Add a DiscardableMemoryAllocator to html_viewer.
Recent changes to skia now require this object, and crash if this
factory isn't supplied. We create the minimal implementation to get a
Google search working in html_viewer. This allocator will free
unlocked chunks once it has reached its memory threshold.

We're starting with a soft limit of 20 megabytes. We'll adjust this
based on observed performance.

Review URL: https://codereview.chromium.org/1016493002

Cr-Commit-Position: refs/heads/master@{#321461}


CrOS-Libchrome-Original-Commit: 5f32298dbf36a4c48ef02d18b6acb1f2be759195