[tracing] child-process-side impl for inter-process memory dumps
This CL introduces the implementation for handling the inter-process
memory dumps on the child-process side.
Conceptually this consists of a simple plumbing in TraceMessageFilter
in order to propagate the request up to the browser's MemoryDumpManager,
as the child processes' MDM play a very passive role in the inter-process
coordination game.
In practice, some extra care is required to match the lifetime of MDM
MDM (long) and ChildTraceMessageFilter (shorter) in a thread-safe
manner, which requires the introduction of the proxy class herein
called ChildMemoryDumpManagerDelegateImpl.

More context and design doc are available in the attached BUG.

Reason for NOTRY: all but one bots are green. I started goo.gl/tBX5f1
on infra-dev about the failure and it seems due to a bot out of space.
The change has been sitting for 4+ hours and there are other CLs blocked
on this.

BUG=462930,474973
NOTRY=true

Review URL: https://codereview.chromium.org/1039963003

Cr-Commit-Position: refs/heads/master@{#324229}


CrOS-Libchrome-Original-Commit: f2f5ca37b0b57ffcc91abad73b6c81e82e0c2511