[tracing] Simplify design of MemoryAllocatorDump
This is a follow-up to crrev.com/1084213003 (which introduces the
MemoryAllocatorAttributesTypes) and crrev.com/1093953002 (which
introduces the shared MemoryDumpSessionState).
The design of MemoryAllocatorDump is simplified, as follows:
MemoryAllocatorDump(s) are uniquely identified witin a
ProcessMemoryDump by an allocator_name and a heap_name.
The former acts as a type, and determines the types
of the attributes. The latter defines the rest of the dump name.
As far as the trace is concerned, the allocator dumps will be
identified the trace by an absolute path, that is:
allocator_name/heap/subheap.

BUG=466141

Review URL: https://codereview.chromium.org/1095003002

Cr-Commit-Position: refs/heads/master@{#325994}


CrOS-Libchrome-Original-Commit: ba8859465a0ff1a6b9eb63f9ed5d4cd8bba69cfb