Append input event type as trace event's name
Previously to show input event's type in the InputLatency
trace event, we issue an async step event with the
input event type as the step event's name. However, trace
viewer's UI for showing async step event is confusing that
the step event is stacked under its parent async event.
E.g. the InputLatency for a GestureScrollUpdate trace event
will be shown as an aysnc trace event with name "InputLatency"
and then a separate async step event with name "GestureScrollUpdate".

To make it cleaner, this CL removes the async step event, and
append the event type to the name of the InputLatency's async
event. So for the above example, there will only be one async
trace event with name "InputLatency::GestureScrollUpdate".

BUG=472121

Review URL: https://codereview.chromium.org/1089243002

Cr-Commit-Position: refs/heads/master@{#326410}


CrOS-Libchrome-Original-Commit: 7f9fd6071aeb387d5c98ca839ac54af555ba2c34