Add a PrefRegistrationFlags::LOSSY_PREF flag for writing lossy prefs
This adds a registration flag for marking a pref as "lossy". There is no strict
time guarantee on when a lossy pref will be persisted to permanent storage when
it is modified.

WriteablePrefStore::PrefWriteFlags are introduced to be used to change the way
that prefs get written into a WriteablePrefStore.

The plumbing to pass registration flags down from PrefService into the
WriteablePrefStore and convert them into PrefWriteFlags is also added here.
The code which actually makes the pref behave in a lossy way has not been
added yet.

BUG=476800

Review URL: https://codereview.chromium.org/1092223004

Cr-Commit-Position: refs/heads/master@{#328481}


CrOS-Libchrome-Original-Commit: 76de1af44ac758e104f2fea2f045f0274aa42221