Fix sparse histograms logging counts of 0.
The way the serialization works, is that a snapshot of the current histogram is
taken. For a sparse histogram, this is done via a SampleMap. Then, the previous
delta (if any) is subtracted, in HistogramSnapshotManager::PrepareDelta().

Then, the snapshot is returned. Iteration through the snapshot is done through
its Iterator() impl.

The SampleVectorIterator has logic to skip empty buckets (since they are very
much the norm), but SampleMapIterator didn't - since they were not the norm
except when there was a delta resulting in a 0 value after subtraction.

This CL fixes that later case and results in a data size reduction of ~1% of
histogram data, according to server-side stats.

BUG=491966

Review URL: https://codereview.chromium.org/1152333006

Cr-Commit-Position: refs/heads/master@{#331598}


CrOS-Libchrome-Original-Commit: ec1eac4fd48cfcc20a3c4f8001a88090beb0a38c