Adding peak resident set size to memory tracing.
This CL adds the peak resident set size in the process runtime, from
the process metrics to the traces. This support is only available in
linux and windows. In addition to the peak, in linux it is possible to
reset the peak rss to current value and this lets us get the peak rss in
the given tracing interval.

BUG=487961

Review URL: https://codereview.chromium.org/1165943002

Cr-Commit-Position: refs/heads/master@{#333276}


CrOS-Libchrome-Original-Commit: 4a3087d6ec3832632acbc8bf96e1aae243830831