Initial cut at exposing frame tree to clients
This is just enough to show the shape of things, with lots of TODOs to
make it real. Here's the general flow (similar to VM):

// When a client in the frame tree is connected to by way of the ViewManager a
// FrameTreeClient is obtained (from the ServiceProvider interface request
// passed in ViewManager::OnEmbed()). The FrameTreeClient is told the frame
// tree (by way of OnConnection()), which allows the client to use other
// frames in the tree (assuming the client has the appropriate permissions).
//
// frame_ids are the same as views ids. This means that when a client creates
// a new view to be part of the frame tree it immediately knows the id to use
// for FrameTreeServer calls.

An alternative is to let the client request FrameTree, set itself as
the FrameTreeClient and then have the server call embed. I kind of
like the flow I have here as it's easier to test.

Also, I'm going to with notifying all clients on every change. We can
change that later on, but this seems simplest for now.

I need to refactor a bunch of things both in the mandoline side and
the html_viewer side to deal with this.

BUG=490221
TEST=covered by tests
R=ben@chromium.org, fsamuel@chromium.org

Review URL: https://codereview.chromium.org/1166723003

Cr-Commit-Position: refs/heads/master@{#333329}


CrOS-Libchrome-Original-Commit: db3ebc1f3e4059f273e8022b8c2d4eabed4e91cf