[tracing] Fix a race when enabling both memory-infra and toplevel.flow
This is a follow-up to crrev.com/1191473004. The general problem is
that we have to ensure that the ThreadLocalEventBuffer(s) are
initialized and have registered their memory dump providers
before memory tracing happens. Otherwise, the act of creating a memory
dump can end up initializing and registering the thread-local dump
provider while the MemoryDumpManager is already locked.

The issue in crbug.com/504311 is that when toplevel.flow is enabled
the act of posting a task creates a trace-event. That happens to be the
very first operation the MemoryDumpManager does OnTraceLogEnabled().
The fix is trivial: initialize the ThreadLocalEventBuffer before
taking the lock in OnTraceLogEnabled.

BUG=504311

Review URL: https://codereview.chromium.org/1207253002

Cr-Commit-Position: refs/heads/master@{#336209}


CrOS-Libchrome-Original-Commit: c766e527bd875dd56567cd8f4915ebcc83cb8dec