Mandoline: Move ContentHandlerConnection into separate file
ContentHandlerConnection will serve as the equivalent of SiteInstance in Chromium.

In the future, ContentHandlerConnection will represent an Application instance for a particular origin/site.

As the complexity of this object will grow soon, it makes sense to move it out as a first step.

ContentHandlerConnection now also uses the same lifetime management idiom used in ApplicationConnection and elsewhere to help guard against reentrant calls during destruction and double deletes. This also makes it easier to unit test this class in the future as more functionality is added.

BUG=496937

Review URL: https://codereview.chromium.org/1228743002

Cr-Commit-Position: refs/heads/master@{#338173}


CrOS-Libchrome-Original-Commit: 09c726e468cf0e5e0c9a2d39888c63561d5838f0