Fix apptests on Android.
The problem was that r338299 moved Context::Init call to run on the main thread on Android, just like it does on desktop, to avoid ApplicationManager being called on multiple threads (which leads to WeakPtr wrong-thread dchecks). This fixed Mandoline on Android, but caused problems with mojo_shell on Android. The reason was that ConfigureAndroidServices wasn't being called before Init.

The fix is to make mojo_shell on Android similar to Mandoline on Android, and mojo_shell and Mandoline on desktop, by having ApplicationManager only be accessed on the main thread. This fixes the weak ptr asserts and also shares more code paths.

BUG=497394

Review URL: https://codereview.chromium.org/1236743002

Cr-Commit-Position: refs/heads/master@{#338561}


CrOS-Libchrome-Original-Commit: f8341ea0704df6b503eba8d90b24456d0ebd3c3e