Stack sampling profiler: re-enable stack collection for stacks terminated by leaf functions
Collect the 5% of leaf-terminated stacks that we're currently missing.
Avoids crashes we saw previously on unwinding frames in injected third
party modules with functions not adhering to the Microsoft x64 calling
conventions, by blacklisting leaf unwinding for these modules when we
first see a bad frame.

With this change there's still a very small possibility of crashing if a
stack happens to start with one of the bad third party functions.  This
change will allow us to collect crash data on this case and inform a
decision about appropriate heuristics to deal with it.

BUG=516402

Review URL: https://codereview.chromium.org/1263403002

Cr-Commit-Position: refs/heads/master@{#345012}


CrOS-Libchrome-Original-Commit: 863f84165da958298516064732d4d5b62ad7eb59