Revert of Move fetching logic out of ApplicationManager, eliminate url mappings. (patchset #12 id:220001 of https://codereview.chromium.org/1342503003/ )
Reason for revert:
Broke the gn build:
ERROR at //mojo/fetcher/about_fetcher.h:8:11: Include not allowed.
#include "mojo/shell/fetcher.h"
          ^-------------------
It is not in any dependency of
  //mojo/fetcher:fetcher
The include file is in the target(s):
  //mojo/shell:shell
which should somehow be reachable.
___________________
ERROR at //mojo/fetcher/base_application_fetcher.cc:12:11: Include not allowed.
#include "mojo/shell/application_manager.h"
          ^-------------------------------
It is not in any dependency of
  //mojo/fetcher:fetcher
The include file is in the target(s):
  //mojo/shell:shell
which should somehow be reachable.
...

https://build.chromium.org/p/chromium.linux/builders/Linux%20GN/builds/31954/steps/gn_check/logs/stdio

Original issue's description:
> Move fetching logic out of ApplicationManager, eliminate url mappings.
>
> BUG=
>
> Committed: https://crrev.com/7134773ceeef31531b84818158902e6d7c74f110
> Cr-Commit-Position: refs/heads/master@{#348659}

TBR=sky@chromium.org,ben@chromium.org
NOPRESUBMIT=true
NOTREECHECKS=true
NOTRY=true
BUG=

Review URL: https://codereview.chromium.org/1343823002

Cr-Commit-Position: refs/heads/master@{#348672}


CrOS-Libchrome-Original-Commit: ec1a9cfea91141cb9b4c0ebe18208afc08caac81