Overhaul Mandoline event transport code
In the course of its development, Mandoline's event code has diverged away
from event payloads that are a good fit with the current state of blink's
input handling logic. This CL begins a process of updating the event code
with an extended wheel interface capable of supporting the standards track
wheel event and revised pointer events that will be extended in a future
CL with support for the pointer events specification.

BUG=524828

Review URL: https://codereview.chromium.org/1313353010

Cr-Commit-Position: refs/heads/master@{#349015}


CrOS-Libchrome-Original-Commit: 895c6c35457b7fe13da31e862cd50eb537d5e7e6