Reland #2 "ipc: Add a new field num_brokered_attachments to the message header."
This original version of this CL is causing an unusual crash in Canary. This CL
adds a message verifier to the ipc code so that it can dynamically verify the
contents of this message that is being corrupted. This CL also verifies the
message at several different points in the dispatch process. This will help
narrow down the range of code that is corrupting the message.

I expect this CL to cause ~100 crashes in the next Chrome Canary. I intend to
revert this CL after a single Canary release.

BUG=527588

Review URL: https://codereview.chromium.org/1334593002

Cr-Commit-Position: refs/heads/master@{#349056}


CrOS-Libchrome-Original-Commit: a2e71be46dc4bdcbb544db479680f65a390ae8f3