Check for CloseHandle failures even when not debugging
Bug 524267 was a handle closing failure that only showed up when running
renderer processes under a debugger at startup, so it was not discovered
for a while.

Similarly, bug 520305 is a long-standing base_unittests bug that only
shows up under a debugger.

This change fixes 520305 and checks for many handle closing failures
always so that subsequent bugs of this type will be detected
immediately.

Most of the CloseHandle calls in base are now checked.

This replaces the uncommitted https://codereview.chromium.org/1343873002/

R=grt@chromium.org,rvargas@chromium.org,dalecurtis@chromium.org
BUG=524267,520305

Review URL: https://codereview.chromium.org/1350493002

Cr-Commit-Position: refs/heads/master@{#349333}


CrOS-Libchrome-Original-Commit: 9ae49a753d07f5a9266a63a89e7336d774f3fe37