[Tracing] Add hook to PartitionAlloc for heap profiling
This modifies PartitionAlloc to call a function on allocation and free,
which will be used for heap profiling. Care has been taken to impact the
fast path as little as possible. https://crbug.com/530502 suggests that
there is no significant performance impact.

This is part of the heap profiler in chrome://tracing.

BUG=524631

Review URL: https://codereview.chromium.org/1391933004

Cr-Commit-Position: refs/heads/master@{#355304}


CrOS-Libchrome-Original-Commit: 65d38472921ad574623410efdfbae840b76fd43e