[Tracing] Add support for adding metadata to the trace log
This adds an |AddMetadataEvent| method to |TraceLog|. It takes a name
(for the metadata event) and args that will go into the "args"
dictionary. When the trace is flushed, |AppendAsTraceFormat| is called
on the values in the args. This enables generating the actual data at
the end of the trace.

This method will be used to add global data to the trace. This saves
space in the trace json because trace events can reference the global
data instead of including the same data inline on every event. Even
though global data is not strictly metadata, the existing metadata
events can support this use case, and re-naming metadata events is not
worth the effort at this point.

This includes a few drive-by include-what-you-use fixes in trace_log.h.

BUG=542702

Review URL: https://codereview.chromium.org/1421053002

Cr-Commit-Position: refs/heads/master@{#356292}


CrOS-Libchrome-Original-Commit: 8d8d218c97ff98a8ffc38d9c658dd04218a4cd37