[Tracing] Fix DCHECK and stack underflow check order
The DCHECK that checks proper nesting of trace events in the allocation
context pseudo stack would dereference the top of stack before it was
checked that the stack is not empty. A pop on an empty pseudo stack can
happen when capturing is enabled when a trace event is in scope.

Review URL: https://codereview.chromium.org/1424223002

Cr-Commit-Position: refs/heads/master@{#356585}


CrOS-Libchrome-Original-Commit: 2e817279585abd3a5eb9574e16f727bc01fd9ec5