Fix race condition in new Mojo EDK exposed by Mac ASAN bots.
The problem is that there are races between when the RawChannel used by ChildBrokerHost notices that the process died compared to the RoutedRawChannel owned by BrokerState. Rather than fix these in BrokerState, this changes removes the RoutedRawChannel owned by BrokerState and instead shares the channel owned by ChildBrokerHost. This has a side benefit of simplifying the code and reducing the number of channels.

BUG=561803

Review URL: https://codereview.chromium.org/1521263002

Cr-Commit-Position: refs/heads/master@{#365083}


CrOS-Libchrome-Original-Commit: 379e90dbc9c051925d672e1b4f7bddf9b43e3af1