Lazily allocate the holder in mojo::internal::SharedData<>.
This makes it possible to create a SharedData on one thread, but
initialise it on another. This solves the issue of null mojo::Callbacks
being bound to the creation thread, preventing them from being assigned
on another (i.e. such as in uses of Binding<>).

BUG=569694

Review URL: https://codereview.chromium.org/1512223002

Cr-Commit-Position: refs/heads/master@{#365764}


CrOS-Libchrome-Original-Commit: b3057bedfb2584157e5e0da9a4936c2c80baf72c