Add run-time CHECK to smoke-test allocator overrides
As part of the refactoring work on the allocator, it emerged that it
would be nice to have a test that ensures that we don't accidentally
break things along the way and not detected that.
This CL adds a runtime CHECK() that verifies that:
 - On Windows (non-component build) the shim layer has been
   initialized. Concretely this checks that chrome called the shimmed
   _heap_init() method and not the one from libcmt.
 - On Linux desktop, the malloc symbols are being overridden by tcmalloc
   (only when USE_TCMALLOC is defined).

The rationale of this test is: if _heap_init was shimmed there are
very good chances that malloc (& friends) were shimmed as well.
Likewise on Linux for mallopt() <-> malloc().

BUG=564618

Review URL: https://codereview.chromium.org/1577883002

Cr-Commit-Position: refs/heads/master@{#369007}


CrOS-Libchrome-Original-Commit: be9d6fce01061afde845fd7aa2335e5757d6a421