Create "persistent memory allocator" for persisting and sharing objects.
This allocator reserves memory within a specific block that itself can
be persisted and/or shared between processes by various means.

It is lock-free for operation on all platforms (even Android which
does not support inter-process locks) and self-checking to deal
with security concerns where not all processes accessing the memory
are fully trustworthy.

BUG=546019

Review URL: https://codereview.chromium.org/1410213004

Cr-Commit-Position: refs/heads/master@{#370377}


CrOS-Libchrome-Original-Commit: 34ae4983d4a24f5136bf8fda7b618842920962b0