Enable attachment brokering of SharedMemoryHandle on Windows.
Whenever a SharedMemoryHandle associated with the current process is passed as a
parameter to a Chrome IPC message, the object is automatically brokered into the
destination process. This means it is no longer necessary to have a constructor
of SharedMemoryHandle that takes a ProcessHandle, since the SharedMemoryHandle
is always associated with the current process.

BUG=493414

Committed: https://crrev.com/7ceb8901740c35cd8ebfb92739efaf2ee441391e
Cr-Commit-Position: refs/heads/master@{#370193}

Review URL: https://codereview.chromium.org/1588683003

Cr-Commit-Position: refs/heads/master@{#373174}


CrOS-Libchrome-Original-Commit: 0da942591fbe276f33c484e8e1e55e069b3c05fb