Clean up public interface of AttachmentBrokerUnprivileged.
In the old interface, a static factory method returns a scoped_ptr, and the
caller had to manage the lifetime. Since this is a global object with minimal
memory footprint, and is required to outlive every IPC::Channel, it's much
easier for the global to never be destroyed. This also matches the interface for
AttachmentBrokerPrivileged.

BUG=584297

Review URL: https://codereview.chromium.org/1679763002

Cr-Commit-Position: refs/heads/master@{#374759}


CrOS-Libchrome-Original-Commit: d00f5b6987750d39a5cc4859ee19c9ca89ed0cd9