Mojo C++ bindings: remove the concept of "GenericInterface".
Previously, if a variant Foo1 was generated for mojom interface Foo. Foo1::GenericInterface was defined as a typedef of the non-variant version Foo.
This required the non-variant version to be generated and depended on by all variants.
The purpose was to make different variants use the same [Associated]Interface{PtrInfo,Request} types.

In fact, this is not very useful in real-world use cases. It should be very rare that users need to deal with multiple variants of the same mojom interface in a single same app.

This change removes GenericInterface and makes [Associated]Interface{PtrInfo, Request} directly use Foo1 instead of Foo1::GenericInterface.

BUG=590329

Review URL: https://codereview.chromium.org/1775613003

Cr-Commit-Position: refs/heads/master@{#379970}


CrOS-Libchrome-Original-Commit: b92c3dd75843e7603892e997a80dcdba5102ce07