Add histograms to compare GetVersionEx() with VerQueryValue() of kernel32
Checking VerQueryValue() of kernel32 reports the "real" OS, rather than
the potentially shimmed one that GetVersionEx() reports.

Normally it's better to use GetVersionEx() because that'll determine the
APIs that are available and how they behave. However, we'd like to know
if there are a substantial percentage of users in compatibility mode, as
there have been complaints of users seeing the "XP and Vista are no
longer supported" infobar on Windows 7.

R=asvitkine@chromium.org, robliao@chromium.org, wfh@chromium.org, nick@chromium.org
BUG=581499

Review URL: https://codereview.chromium.org/1784623003

Cr-Commit-Position: refs/heads/master@{#380793}


CrOS-Libchrome-Original-Commit: d68b1e1c89ab3d8055046883b1bf726084673959