TaskScheduler [4/9] Priority Queue
This change is a subset of https://codereview.chromium.org/1698183005/

A PriorityQueue holds Sequences of Tasks. It supports Push, Pop and
Peek operations through a Transaction object.

A SequenceSortKey must be provided to push a Sequence into a
PriorityQueue. Sequences are sorted according to their SequenceSortKey.
The SequenceSortKey of a Sequence never changes while it is in the
PriorityQueue (even if Tasks are pushed/popped from the Sequence).

BUG=553459

Review URL: https://codereview.chromium.org/1709713002

Cr-Commit-Position: refs/heads/master@{#382115}


CrOS-Libchrome-Original-Commit: 6477fbded94522e26875bbf405710060d01ea971