tracing: add dump provider for malloc heap profiler
Now that the allocator shim is landing (the base API landed in
crre.com/1675143004, Linux support in crrev.com/1781573002) we
can use that to implement the malloc heap profiler on top of that.
This CL leverages the new shim API to collect information about
malloc/free and injecting that into the heap profiler, leveraging
the same infrastructure we are already using for PartitionAlloc
and BlinkGC (oilpan).

Next steps
----------
Currently this heap profiler reports also the memory used by the
tracing subsystem itself, which is not desirable. Will come up
with some scoped suppression semantic to bypass the reporting.

BUG=550886
TBR=haraken

Review URL: https://codereview.chromium.org/1675183006

Cr-Commit-Position: refs/heads/master@{#382505}


CrOS-Libchrome-Original-Commit: 14c8b52dac1285bbf23514d05e6109aa7edc427f