TaskScheduler [5/9] Task Tracker
This change is a subset of https://codereview.chromium.org/1698183005/

All tasks go through the scheduler's TaskTracker when they are posted and
when they are executed. The TaskTracker enforces shutdown semantics and takes
care of tracing and profiling.

We have an alternate implementation which uses atomic operations to reduce lock
contention. However, we prefer to check this simple implementation in first and
bring the atomic operations in an incremental CL.

BUG=553459

Review URL: https://codereview.chromium.org/1705943002

Cr-Commit-Position: refs/heads/master@{#383749}


CrOS-Libchrome-Original-Commit: 0fc7a6669c12c2ef2d1ccbd496232b7c65f6c81c