TaskScheduler: Pop a Task from its Sequence from SchedulerWorkerThread.
Previously, we planned to pop a Task from its Sequence inside the
RanTaskFromSequence() method of SchedulerWorkerThread::Delegate. This
method would also have pushed the Sequence to the appropriate
PriorityQueue if it wasn't empty after the pop.

We now believe that it is more natural to pop the Task from its Sequence
from SchedulerWorkerThread, right after it runs. If the Sequence isn't
empty after the pop, SchedulerWorkerThread will call
SchedulerWorkerThread::Delegate::EnqueueSequence(). It won't notify  its
delegate if the Sequence is empty after the pop.

BUG=553459

Review URL: https://codereview.chromium.org/1862243005

Cr-Commit-Position: refs/heads/master@{#386171}


CrOS-Libchrome-Original-Commit: d314e56ae71bfb19fa9d1745055936cc870652e1