[tracing] Ignore tracing allocations in heap profiler
This CL removes the memory usage of tracing from the heap profiler.
Adds support for an ignore event which clears the stack frame and
type for the allocations happening in that scope and reset them to
"overhead". We do not make it null because it will show under "<other>"
or "[unknown]".
Adds the ignore events to major functions that allocate memory for
tracing. The total is usually 500Kib more than the total in "tracing"
dump.

BUG=604689

Review URL: https://codereview.chromium.org/1900223003

Cr-Commit-Position: refs/heads/master@{#389159}


CrOS-Libchrome-Original-Commit: 04d1750c6721a424ae45e7e28f71dba69f1789db