[iOS] Remove dependency third_party/WebKit on iOS.
Chrome on iOS now requires mojo for WebUI which has an optional dependency
on third_party/WebKit. Chrome on iOS cannot depends on third_party/WebKit,
so only enable those optional dependencies if target OS is not iOS.

Fixes the following build failures when building with clang Xcode (which
is only done by downstream release bots as we do not have capacity to run
it on each commit):
  error: unable to load plugin 'third_party/llvm-build/Release+Asserts/lib/libBlinkGCPlugin.dylib': 'dlopen(third_party/llvm-build/Release+Asserts/lib/libBlinkGCPlugin.dylib, 9): Symbol not found: __ZN4llvm8RegistryIN5clang15PluginASTActionEE4HeadE
    Referenced from: third_party/llvm-build/Release+Asserts/lib/libBlinkGCPlugin.dylib

To prevent further dependencies on third_party/WebKit to creep in change
third_party/WebKit/Source/{config.gyp,wtf/wtf.gyp} to not define any
target when target OS is iOS (in order to generate errors when running
"gclient runhooks").

This is similar to corresponding GN changes in issue
https://codereview.chromium.org/2004743002/.

BUG=616750

Review-Url: https://codereview.chromium.org/2036683002
Cr-Commit-Position: refs/heads/master@{#398272}


CrOS-Libchrome-Original-Commit: aa9125d2a9f1417d069a99d178f86784f8210215