Use generated list of java sources as Android Lint input.
We are already using GN-written list of java sources as an input for
javac and emma_instrumentation actions instead of passing all files via
command line. This CL switches lint action to use the same file. While
it isn't needed for Android Studio support which is the original reason
this source list was added, passing sources as list improves consistency
and avoids hitting a command-line limit imposed by OS (which is large
but still reachable even on Linux).

BUG=none

Review-Url: https://codereview.chromium.org/2159173002
Cr-Commit-Position: refs/heads/master@{#406271}


CrOS-Libchrome-Original-Commit: b54d076261882c53f0dd9c9e1707debe0feafb2a