Tracing pre-filtering
This adds functionality to insert filtering predicates in-between calls to add trace events, and the actual trace buffer. These predicates can be used to implement custom logic for whether or not events should actually be added or not, in addition to the normal category checks. Events can also be modified, if need be.

The initial added usecase for this is to be able to add an additional level of filtering to reduce sizes of traces to the bare minimum, for use with perf waterfall TBMv2 metrics, and Slow Reports.

Review-Url: https://codereview.chromium.org/1923533004
Cr-Commit-Position: refs/heads/master@{#412594}


CrOS-Libchrome-Original-Commit: ca609aea2cd3f4c25766f1f7d2715f4c1e54e296