Mojo EDK: Add tests to exercise races in handle transit
Data pipe producer handles may change status while in transit
and the receiving end may never be notified of this change unless
they call MojoWait on the handle.

This fixes that, and adds test coverage to verify that the problem
does not exist and does not regress for data pipe consumers (fixed by
http://crrev.com/411707) or message pipe handles (don't currently
have this problem) either.

BUG=639363
R=yzshen@chromium.org

Review-Url: https://codereview.chromium.org/2259363002
Cr-Commit-Position: refs/heads/master@{#413356}


CrOS-Libchrome-Original-Commit: 7cd99cc18b960e7d59109f16c53cad49ac09e7c4