Add real_path option to LaunchProcess
When specified, real_path will be used as the location of the executable
to launch instead of argv[0]. This is useful when a custom argv[0] must
be specified.

This is needed by Chrome Remote desktop to start a login shell, which
requires invoking the shell with an argv[0] starting with a '-'
character (e.g., "-bash").

Review-Url: https://codereview.chromium.org/2338133005
Cr-Commit-Position: refs/heads/master@{#422457}


CrOS-Libchrome-Original-Commit: 732f03d294cead0635f02d24a3fa656416f58c25