tracing: remove sampling state profiler
As per discussion in [1] the tracing sampling state profiler is not used anymore.
The coverage of the TRACE_EVENT_SCOPED_SAMPLING_STATE macros became
poor over time and new alternatives came out (e.g., V8 sampling profiler).
The overall is sensible but should be re-approached in a more sustainable way.
More specifically, the new trace event filters could be leveraged to build a
sampling profiler on top of the existing TRACE_EVENT macros.

[1] https://groups.google.com/a/chromium.org/d/msg/tracing/E3K6qF1YMI8/wapmBei2AQAJ

BUG=

Review-Url: https://codereview.chromium.org/2406703002
Cr-Commit-Position: refs/heads/master@{#424212}


CrOS-Libchrome-Original-Commit: 7bc681e5c39bda0926e77961273916dc90530558