VideoCapture: more migration IPC-->mojo, part 6
This CL does the migration of the last Vid Capture IPC
message to mojo, namely OnBufferCreated().  This
message shared a SharedMemoryHandle, which is now
send using mojo::{Wrap,Unwrap}SharedMemoryHandle().
This needs a few small changes elsewhere in the code,
in particular there is no need to ShareToProcess() a given
ShMem (and this in turn makes us think that the whole
data-sending can be simplified in later CLs).

Following suggestion in PS1, VideoCaptureBufferTracker()
is changed to
mojo::ScopedSharedBufferHandle GetHandleForTransit();,
which causes changes in a few other classes;
- SharedMemoryBufferTracker, MojoSharedMemoryBufferTracker
- VideoCaptureBufferPool & Impl,
- VideoCaptureController (no need for render process id).
(all these changes are in services/video_capture/ and
content/browser/renderer_host/media/)

VideoCaptureHost is still a BrowserMessageFilter,
and VideoCaptureMessageFilter is not completely
gone yet, but will disappear soon -- I didn't want to
mix the changes in the ShMem management with
other changes just in case.

BUG=651897
TEST=local capture, {content,capture,video_capture}_unittests

Review-Url: https://chromiumcodereview.appspot.com/2410383002
Cr-Commit-Position: refs/heads/master@{#426500}


CrOS-Libchrome-Original-Commit: 0abf0b0b42e9e974d78b6f13a2ed81afdd668287