IPC: Relax thread constraints on Channel-associated send
This removes a DCHECK from ChannelAssociatedGroupController, allowing
messages to be sent from arbitrary threads rather than only from the
IO thread or proxy thread.

While it is still unsafe to bind associated endpoints on non-proxy
non-IO threads, this allows control messages to be transmitted
safely from any thread, so for example an AssociatedPtrInfo can safely
notify the request side of closure if it's destroyed on an arbitrary
thread.

BUG=663046
R=yzshen@chromium.org

Review-Url: https://codereview.chromium.org/2481963002
Cr-Commit-Position: refs/heads/master@{#430705}


CrOS-Libchrome-Original-Commit: becd3f741efe619c34bb569fd4ab18d85da45d6f