Modernize TimerTests prior to https://codereview.chromium.org/2491613004
https://codereview.chromium.org/2491613004 failed TimerTest.OneShotTimer_CustomTaskRunner
with the old tests per lacking a TaskRunnerHandle instance. Fixing that
led to a bunch of cleanup. Decoupling from that CL to make it clear that the
new tests pass in the old world and that the new world then passes the tests
without tweaking them.

Mostly moving towards RunLoop/TaskRunners/WaitableEvents to ease a few things.

Also removed incorrect DCHECK in Timer::PostNewScheduledTask() which was tripped
by the new version of TimerTest.OneShotTimer_CustomTaskRunner

BUG=587199, 552633

Review-Url: https://codereview.chromium.org/2514293004
Cr-Commit-Position: refs/heads/master@{#434764}


CrOS-Libchrome-Original-Commit: 175490c1043e6c6ce85d42eaeb1dcfb34b2c9267