[memory-infra] Add support for polling memory totals in MemoryDumpManager
This adds new method in MemoryDumpProvider interface PollFastMemoryTotal
which is used for polling memory totals in process. The dump provider
has to enable this feature while registration.
For Simplicity of polling logic the dump providers supporting polling
must not have task runner affinity. Polling will be done on dump thread.

BUG=607533

Review-Url: https://codereview.chromium.org/2537363003
Cr-Commit-Position: refs/heads/master@{#439390}


CrOS-Libchrome-Original-Commit: 4460b9b6393448adc011207ae63380a70ddba369