Mojo: Introduce InterfaceRequest ctor that takes in InterfacePtr*
In https://groups.google.com/a/chromium.org/forum/#!topic/chromium-mojo/PUDEadGMijM,

we decided to introduce an InterfaceRequest constructor that takes in an
InterfacePtr* and binds to it on a new pipe (i.e., the same functionality
as mojo::MakeRequest(InterfacePtr*)). The primary use case is to replace
lines like:

FooRequest request = mojo::MakeRequest(&foo_ptr);

with:

FooRequest request(&foo_ptr);

This CL does the following:

(1) Introduces the new constructor.
(2) Changes mojo::MakeRequest(InterfacePtr*) to be implemented via this
    constructor.
(3) Performs the above-mentioned conversion across the codebase.

Note that it is not possible to use "auto" with this constructor; replacing
FooRequest with auto in the above construction will cause the compiler to
infer the type to be FooPtr*.

A followup will make a similar change to AssociatedInterfaceRequest.

CQ_INCLUDE_TRYBOTS=master.tryserver.chromium.linux:linux_site_isolation
TBR=darin

Review-Url: https://codereview.chromium.org/2592623002
Cr-Commit-Position: refs/heads/master@{#440084}


CrOS-Libchrome-Original-Commit: c17a15dfa5b357271cd283523a868688415ef919