Tag some of Mojo heap allocations for the heap profiler.
First step in tagging heap allocations for Mojo as to provide context
for the heap profiler.
Allocations done by mojo::Watcher are now tagged with the master
interface name if available, otherwise with the location where the
Watcher was allocated.
Non Mojo IPCs are tagged as "IPC Channel" at this point.

BUG=676682
TBR=sergeyu

Review-Url: https://codereview.chromium.org/2623263005
Cr-Commit-Position: refs/heads/master@{#446435}


CrOS-Libchrome-Original-Commit: 2207af1a2e5f829ea98198d87802872781df60b8