Fix clang static analyzer false positives due to DCHECK.
DCHECK causes many false positives because its implementation defeat some
of the assumption that the static analyzer makes. In particular, DCHECK
that a pointer is not null cause the static analyzer to report that all
usage of the pointer may be null.

This can be fixed by changing the implementation of the DCHECK macro to
introduce a call to a function marked with the attribute analyzer_noreturn
when the condition is false as suggested in the clang documentation.

BUG=648210

Review-Url: https://codereview.chromium.org/2625443003
Cr-Commit-Position: refs/heads/master@{#447080}


CrOS-Libchrome-Original-Commit: 202539e475b6398e7181706d44b43a22f0711227