AssociatedGroupController impls: reduce message pipes created for sync message signalling.
Previously, we always create a message pipe for sync message signalling when
SignalSyncMessageEvent() is called. And this method is always called when peer
endpoint is closed. Most of time, this message pipe is not used.

BUG=None
TBR=rockot@chromium.org
(rockot has LGed and is an owner of all files. But somehow presubmit check thought ipc_mojo_bootstrap.cc needed an OWNER review.)

Review-Url: https://codereview.chromium.org/2718253005
Cr-Commit-Position: refs/heads/master@{#453714}


CrOS-Libchrome-Original-Commit: e25b5d5d3092a9ccd7c4b146d787cb8a403d9ec4