Added regression test for base::JoinString given a null StringPiece.
This tests against what was previously undefined behaviour (passing a
null pointer to the std::string constructor), introduced in r452432 and
inadvertently fixed in r454189. These tests won't necessarily catch
future undefined behaviour (because the standard library doesn't
actually seem to mind being given a null pointer) but we stand a better
chance of finding it if there are tests.

BUG=698073

Review-Url: https://codereview.chromium.org/2733473002
Cr-Commit-Position: refs/heads/master@{#454817}


CrOS-Libchrome-Original-Commit: 601ab6b021ca49678a407f4feb66a313318c18aa