base: Introduce SharedMemoryTracker for POSIX (but not macOS)
This CL introduces SharedMemoryTracker to measure memory usage of
SharedMemoy correctly and report them to memory-infra.
SharedMemoryTracker records the memory usage when mmap-ing and
when unmmap-ing, and reports them when OnMemoryDump is called.
This enables to help us know SharedMemory usage correctly.

This is a part of a CL (https://codereview.chromium.org/2535213002/) for
SharedMemory usage measuring.

BUG=604726

Review-Url: https://codereview.chromium.org/2654073002
Cr-Commit-Position: refs/heads/master@{#455406}


CrOS-Libchrome-Original-Commit: 2acea43a7272838a75e4c6b7e5473dd6f9c4aa79