Mojo: Add signal for new data pipe consumer data
Adds a new MOJO_HANDLE_SIGNAL_NEW_DATA_READABLE signal. This signal is
raised on a data pipe consumer handle whenever new data is written into
the pipe, and is held high until the next MojoReadData or
MojoBeginReadData attempt on the handle.

This may act as a more useful signal for data pipe consumers who want to
wait for a minimum amount of data to become available before reading, as
it can be watched independent of general readability so that new work is
only attempted when new data is ready.

BUG=700171

Review-Url: https://codereview.chromium.org/2741353002
Cr-Commit-Position: refs/heads/master@{#456450}


CrOS-Libchrome-Original-Commit: e9032c783d6a470dcd04d3af16cb02d4ee621f7b