android: Java-based launcher thread
A major part of android's ChildProcessLauncher is implemented in Java
so having an easy access to the launcher thread is important.

Refactor JavaHandlerThread in base so that that java part can be created
first, and then connected with native peer after consutrction. This is
needed because the launcher thread can be used before native library is
loaded.

Add LauncherThread in content which is a simple wrapper around
JavaHandlerThread. Then for android, override the launcher thread
message loop. Note this means the launcher thread will no longer be
joined on shutdown, but this is not a problem in practice since android
never does clean shutdowns.

Convert two cases of random background threads in ChildProcessLauncher
to use LauncherThread. There are more in the future.

BUG=701442

Review-Url: https://codereview.chromium.org/2774363003
Cr-Commit-Position: refs/heads/master@{#460509}


CrOS-Libchrome-Original-Commit: 61687ec5b7b5dce1c9c86ae46f552c0c34473555