RedirectNonUINonIOBrowserThreads to TaskScheduler by default on trunk.
The addition of AssertWaitAllowed in BrowserMainLoop::ShutdownThreadsAndCleanUp()
is required because joining is a "wait" operation. It wasn't required
before because PlatformThread::Join (platform_thread_posix.cc) actually still
does AssertIOAllowed() instead of AssertWaitAllowed() for legacy reasons
(and I/O is allowed in shutdown already). But this CL enables the redirection
which uses a WaitableEvent to mimic thread join (ref. BrowserThreadImpl::StopRedirectionOfThreadID():
waiting on the exact same set of tasks it did prior to redirection).

BUG=653916

Review-Url: https://codereview.chromium.org/2690183002
Cr-Commit-Position: refs/heads/master@{#461481}


CrOS-Libchrome-Original-Commit: b6d0c9a06ab573af760763fcd7ecf76ae2c87695