Switch VS 2017 builds to VS 2017 Update 3 Preview 1
This change switches the VS 2017 package to use the first Update 3
Preview.

Packaging was done on a Windows Server 2016 VM, cleanly created for this
purpose.

Compiler was packaged up by downloading the VS 2017 Update 3 Preview 1,
from https://www.visualstudio.com/vs/preview/, and then running this
command (executable name was different):

    vs_professional.exe --add Microsoft.VisualStudio.Workload.NativeDesktop
    --add Microsoft.VisualStudio.Component.VC.ATLMFC --includeRecommended
    --passive

Then the Windows 10.0.14393.0 SDK (10.0.14393.795, to be precise) was
installed, necessary because Chrome currently requires that SDK. This
also installs the x86 and x64 debuggers.

Then statreg.h in the VS install was patched to add constexpr for
clang-cl (but not for VC++, which can't handle it), per this bug:

https://developercommunity.visualstudio.com/content/problem/58907/statregh-doesnt-compile-with-clang-cl-constconstex.html

The patched version is attached to this bug comment:
    https://bugs.chromium.org/p/chromium/issues/detail?id=683729#c113

Finally the packaging script (updated in
https://chromium-review.googlesource.com/516442) was run:

    python depot_tools\win_toolchain\package_from_installed.py 2017 -w 10.0.14393.0

VS 2015 builds are still the default.

R=scottmg@chromium.org
BUG=683729

Review-Url: https://codereview.chromium.org/2904733004
Cr-Commit-Position: refs/heads/master@{#475088}


CrOS-Libchrome-Original-Commit: bc58befabe87e204650c836793204bac887992bd