[Identity Service] Introduce typemaps for a couple core types
We will shortly be adding an interface to obtain access tokens to the
Identity Service. This interface requires a set specifying the requested
scopes and must pass back an error in case of failure. This CL adds
typemaps of the underlying types that OAuth2TokenService uses for this
functionality.

TBR=yzshen@chromium.org

Bug: 683124
Change-Id: Iec1e3eca143e16c5dbdbe34fdb0c5850866f6218
Reviewed-on: https://chromium-review.googlesource.com/514006
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Mihai Sardarescu <msarda@chromium.org>
Cr-Commit-Position: refs/heads/master@{#476974}

CrOS-Libchrome-Original-Commit: 52148d9dec33be3f66dc01010b8309bcb247e64c