[base] Fix macros and comments for C++11
Since the DISALLOW_* macros now use the C++11 syntax "= delete", it
does not matter whether they are used in the public or the private
section of a class definition.
In fact, the style guide recomments putting deleted constructors or
assignment operators in the public section.
Thus, remove the hint to put them in the private: declarations.

Drive-by: Fix the syntax of the deleted copy assignment operator.

R=thakis@chromium.org

Change-Id: I10af0d71c18dc1f3b08e4601ae571ed53d8c71e3
Reviewed-on: https://chromium-review.googlesource.com/578027
Reviewed-by: Nico Weber <thakis@chromium.org>
Commit-Queue: Clemens Hammacher <clemensh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#488172}

CrOS-Libchrome-Original-Commit: cec5fdea187d0edf9491000ba9989129f188e50e