Validate all histograms before and after merging subprocess ones.
This additional instrumentation will tell us if the corruption is
happening as a result of merging subprocess histograms, or whether
things are already corrupt when we get to that point. Note: We've
observed that the corrupt entries are all histograms coming from
child processes.

BUG=744734,736675

Change-Id: Ic772beefa397f387ba905cc8aa3ea73e01dab36b
Reviewed-on: https://chromium-review.googlesource.com/586913
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Commit-Queue: Alexei Svitkine (OOO July28-Aug6) <asvitkine@chromium.org>
Cr-Commit-Position: refs/heads/master@{#489775}

CrOS-Libchrome-Original-Commit: e0a8dda28a72d5bf2ad3c96c426bd373a35944b5