Make mojo_public_bindings_unittests confirm correct sequence, not thread
Two tests, AssociatedInterfaceTest.ThreadSafeAssociatedInterfacePtr and
InterfacePtrTest.ThreadSafeInterfacePointer were checking that their
methods were invoked on the same thread. However, the requirement is
that they're invoked on the same sequence, so change the test to check
for that instead.

(This was exposed running on Fuchsia, but applies to all platforms.)

Bug: 752167
Change-Id: I2f2e81fb2d41878601ff4ab23cb9c9cb0f2c2a8e
Reviewed-on: https://chromium-review.googlesource.com/602484
Reviewed-by: Gabriel Charette <gab@chromium.org>
Reviewed-by: Yuzhu Shen <yzshen@chromium.org>
Commit-Queue: Scott Graham <scottmg@chromium.org>
Cr-Commit-Position: refs/heads/master@{#492628}

CrOS-Libchrome-Original-Commit: 45bbf78dcf9448b9358d51bc2cd06fadeeb992d1