Fix bug where downstream tests do not run
The changes for cronet test apk AndroidManifest and our junit runner
is because the netty jar depended by cronet test apk doesn't include
the classes needed by netty's classes. This would throw exceptions
when we try to list out tests through our junit runner and the runner
try to load all the packages it can find when `-e package` is not
specified.

This is a refactor of the CL 616025 that was reverted because this
issue.

Bug: 755721
Cq-Include-Trybots: master.tryserver.chromium.android:android_cronet_tester
Change-Id: Ic6d2471c94b84d9f5b2a2a5b98a4f1f5ad7067b3
Reviewed-on: https://chromium-review.googlesource.com/629697
Commit-Queue: Yoland Yan <yolandyan@chromium.org>
Reviewed-by: John Budorick <jbudorick@chromium.org>
Reviewed-by: Andrei Kapishnikov <kapishnikov@chromium.org>
Cr-Commit-Position: refs/heads/master@{#497606}

CrOS-Libchrome-Original-Commit: 561eb0e9fc0408e11ac9f1566595252b0e55b551