Support BlockingType::MAY_BLOCK in SchedulerWorkerDelegateImpl.
Initially written by jeffreyhe@:
https://chromium-review.googlesource.com/c/chromium/src/+/633277

This CL adds support for BlockingType::MAY_BLOCK.
SchedulerWorkerDelegateImpl::BlockingScopeEntered() will record a
timestamp of when the worker was blocked. This
timestamp is used to determine whether a thread has been in the scope of
a ScopedBlockingCall long enough that |worker_capacity_| needs to be
increased. When the last idle worker gets woken up, a task will be posted
to the service thread that will cause these timestamps to be periodically
monitored.

In this CL, BlockingType::WILL_BLOCK will continue to have its
|worker_capacity_| increases be done immediately when ScopedBlockingCall
is instantiated.

Bug: 757475
Change-Id: Ic7b5fc8c4d1f77adda48cafee262f3cbeff1711c
Reviewed-on: https://chromium-review.googlesource.com/639330
Commit-Queue: Francois Doray <fdoray@chromium.org>
Reviewed-by: Gabriel Charette <gab@chromium.org>
Reviewed-by: Robert Liao <robliao@chromium.org>
Cr-Commit-Position: refs/heads/master@{#499689}

CrOS-Libchrome-Original-Commit: dcbe7e7c4a00cfb5673e18f568aa2b9e4169b8e8