Mojo C++ MultiplexRouter: use a base::small_map with 1 inlined slot for endpoints.
The common case is that the master interface is the only interface running on
a message pipe (i.e., no associated interfaces). Using a base::small_map
is expected to improve memory usage and performance.

Another purpose of this CL is to see whether the change of memory allocation
behavior has any impact on the (possible) memory corruption bug affecting this
map. (crbug.com/754946)

BUG=754946

Change-Id: I0a0c569550421cb8be6efa505866fba7be16286b
Reviewed-on: https://chromium-review.googlesource.com/665175
Reviewed-by: Ken Rockot <rockot@chromium.org>
Commit-Queue: Yuzhu Shen <yzshen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#501817}

CrOS-Libchrome-Original-Commit: 80f3b831e640316666dfdf430b8efb8ce2c91b23