Properly look up corrupted strings.
Getting the previously freed object at "this" is only useful if the
corruption was found in the "dummy" field.  If in fact it is the
histogram name that has been damaged, we must look up the object from
the string's base address.

About 1/2 of the crashes are due to a corrupted string so this will
effectively double the number of data points to analyze.

Bug: 744734
Change-Id: I9a923b7c416330d906bf2373ced3df08874299aa
Reviewed-on: https://chromium-review.googlesource.com/683734
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Reviewed-by: Ilya Sherman <isherman@chromium.org>
Commit-Queue: Brian White <bcwhite@chromium.org>
Cr-Commit-Position: refs/heads/master@{#504345}

CrOS-Libchrome-Original-Commit: 3239874d6740a9181cbbfcbbdfd0ed58147975d1