Migrate multi-threaded cert verifier from WorkerPool to TaskScheduler.
WorkerPool is being deprecated in favor of TaskScheduler.

NSS functions may reenter //net via extension hooks. If the reentered
code needs to synchronously wait for a task to run but the thread pool
in which that task must run doesn't have enough threads to schedule it,
a deadlock occurs. To prevent that, calls to NSS functions are made
within the scope of a MAY_BLOCK base::ScopedBlockingCall which
increments the thread pool capacity after a short timeout.

Bug: 659191
Change-Id: I2a2719f7c104f1ecf8cac7e4019836920c777261
Reviewed-on: https://chromium-review.googlesource.com/648328
Commit-Queue: Francois Doray <fdoray@chromium.org>
Reviewed-by: Gabriel Charette <gab@chromium.org>
Reviewed-by: Ryan Sleevi <rsleevi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#504439}

CrOS-Libchrome-Original-Commit: 524a99e5ca3aa7df42d743db73fee483f624c1e3