Remove whitespace generated by mojo js template
The validation macros for javascript generate whitespace in the
resultant file.  This change cuts out that whitespace.

This is important so that unmodified versions of these generated files
can be placed into the third_party/WebKit/LayoutTests/external/wpt/
directory which is exported to github.

BUG=509038

Change-Id: Ie09e90935324a87380376ddeaf7d527ea9f9e93b
Reviewed-on: https://chromium-review.googlesource.com/701370
Reviewed-by: Yuzhu Shen <yzshen@chromium.org>
Commit-Queue: Conley Owens <cco3@chromium.org>
Cr-Commit-Position: refs/heads/master@{#506621}

CrOS-Libchrome-Original-Commit: 51a5a000bd005e5c7e428c8d1b5926a7b033745c