Add enable_mojo_tracing build argument.
Implement a build option to insert trace points to the Mojo templates when the
enable_mojo_tracing gn flag is enabled. It's useful when debugging Mojo
comminucation. A large number of trace points are generated, which increases
code size. For this reason this flag is false by default.

It also turns MOJO_TRACE_ENABLED macro to the build flag.

Change-Id: I747f5ec77e0501c391642afa4c751b0a807c434d
Reviewed-on: https://chromium-review.googlesource.com/711842
Reviewed-by: Ken Rockot <rockot@chromium.org>
Commit-Queue: Ken Rockot <rockot@chromium.org>
Cr-Commit-Position: refs/heads/master@{#508405}

CrOS-Libchrome-Original-Commit: 3643a9719e6510ba9da89bd95f11861381607b5a