Record browser process exits in tracker.
The tracker needs to be notified when a process exits but the Process
object doesn't do that (and can't do that) on its own so it needs to
be called where the termination status is fetched.

It still makes sense to pipe this through the Process object just to
keep things in one place and make it easier to be called from other
places where necessary.

Bug: 757946
Change-Id: I9e039bcd143277461e6b8323f1756e03a46f005b
Reviewed-on: https://chromium-review.googlesource.com/750281
Reviewed-by: Mark Mentovai <mark@chromium.org>
Reviewed-by: Antoine Labour <piman@chromium.org>
Commit-Queue: Jamie Madill <jmadill@chromium.org>
Cr-Commit-Position: refs/heads/master@{#513577}

CrOS-Libchrome-Original-Commit: ae2a8b9a9455eede005c731e053e22ab98a0f6d7