Add handlers to crash tests more informatively
The VC++ runtime "helpfully" catches various types of failures (pure
virtual function calls, abort, and invalid parameters) and turns them
into calls to abort. This is good for security but is bad for testing
because it means that no call stack is printed on these "crashes". This
CL changes the behavior so that these become actual crashes, with call
stacks printed by the test framework.

Messages printed (tested in debug builds) before the back trace now
are:

  abort() has been called
  Received fatal exception EXCEPTION_BREAKPOINT
  Backtrace:

  ...\output.cpp(273) : Assertion failed: ("Buffer too small", 0)
  Received fatal exception EXCEPTION_BREAKPOINT
  Backtrace:

  Pure-virtual function call. Terminating.
  Received fatal exception EXCEPTION_BREAKPOINT
  Backtrace:

Bug: 783040
Change-Id: I2f3d666af338e9b52b557ebbfb076b353888d6ff
Reviewed-on: https://chromium-review.googlesource.com/772740
Commit-Queue: Bruce Dawson <brucedawson@chromium.org>
Reviewed-by: Mark Mentovai <mark@chromium.org>
Cr-Commit-Position: refs/heads/master@{#516939}

CrOS-Libchrome-Original-Commit: da276c95c4643ab2ddbbd9a3fe9c84aa16600fa2