android: Log and plot native library code residency.
When enabled with the --log-native-library-residency command line switch,
periodically collects native library residency, and dumps it to a
file. process_residency.py then takes the output and creates a graph from
it. This is used to assess the efficiency of the code ordering logic.

Bug: 746665
Change-Id: I4a21ce4b9c01f8bb255f5997e665ff3dbdcf240a
Reviewed-on: https://chromium-review.googlesource.com/756847
Commit-Queue: Benoit L <lizeb@chromium.org>
Reviewed-by: agrieve <agrieve@chromium.org>
Reviewed-by: Matthew Cary <mattcary@chromium.org>
Reviewed-by: Egor Pasko <pasko@chromium.org>
Cr-Commit-Position: refs/heads/master@{#517385}

CrOS-Libchrome-Original-Commit: 27cc8dfa29dbd26089f8f61846f37b897d3e7f7a