Fix ScopedTaskEnvironmentTest.DelayedTasks to not be racey.
The test created an ASYNC-mode ScopedTaskEnvironment, but then applied
expectations to it that are only valid in QUEUED (synchronous) mode. The
test is fixed to use a QUEUED-mode instance.

A brief delay is also added to the test, to make it more likely to fail
if the behaviour regresses to being asynchronous & racey.

Bug: 789712
TBR: gab
Change-Id: I73808b1633bd7afcb6e0204a4030e16ef323f868
Reviewed-on: https://chromium-review.googlesource.com/804691
Reviewed-by: Wez <wez@chromium.org>
Commit-Queue: Wez <wez@chromium.org>
Cr-Commit-Position: refs/heads/master@{#521221}

CrOS-Libchrome-Original-Commit: f63253ce95cedeeac95cad6951545b872872e229