Start Onion Soup of WebClipboard, by moving shared enums to mojom.
This CL is preparation for onion souping of WebClipboard.

It moves public enum types from WebClipboard into a mojom file
defined in WebKit/common.

The clipboard mojo interface is then changed to use these enums
rather than the type defined in ui/base/clipboard.

Typemaps for clipboard_format and ui::ClipboardType have been
removed as we no longer use these types in the mojom, and
it will make it simpler to move the files around. We may
add type-mapping back in when we've removed WebClipboard.

Code that previously converted from WebClipboard::Buffer to
ui::Clipboard has been moved to the browser side, on the
renderer side we still check that the Buffer type is valid,
but we no longer do the conversion.

The ultimate goal is to replace WebClipboard with the mojo
interface.

Bug: 775830
Change-Id: I2cf4e7fc771dd66c4b6ec95abc7d131686683d52
Reviewed-on: https://chromium-review.googlesource.com/804774
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Stuart Langley <slangley@chromium.org>
Cr-Commit-Position: refs/heads/master@{#521478}

CrOS-Libchrome-Original-Commit: 4f843cc37b3b95af2c382197e4c96997ee95c32a