Switch LaunchWithHandleTransfer test to use action_timeout().
This test times-out on Fuchsia bots, seemingly due to an extremely long
delay in getting around to scheduling the sub-process to execute, some
of the time. Other similar ProcessUtilTests succeed because they wait
indefinitely for the sub-process to do its work; raising the timeout
to action_timeout() provides enough head-room that this test passes.

This also has the advantage that the test's timeouts are now
configurable.

Bug: 793412
Change-Id: I8b43e3cadfe302cf8afd9e835ee4ed1fbcda3c9e
Reviewed-on: https://chromium-review.googlesource.com/828383
Reviewed-by: Scott Graham <scottmg@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/master@{#524639}

CrOS-Libchrome-Original-Commit: 2490b28ceb23d899791b0a23d6d8c59b70570382