viz: Sync destruction for RootCompositorFrameSinkImpl.
For not OOP-D, when destroying a viz::Display there is a sync IPC in
CommandBufferProxyImpl::DisconnectChannel() that ensures the GL
context/surface that draw to the platform window are destroyed. The
platform window is destroyed after that sync IPC which ensures the GPU
doesn't try to swap to a buffer that doesn't exist.

Currently for OOP-D, HostFrameSinkManager sends an async IPC to the GPU
process asking it to destroy the viz::Display. The browser process
destroys the platform window after sending the async IPC. The GPU may
try to swap to a buffer that doesn't exist and error/hang/crash.

Add a sync IPC, only used in the case of root CompositorFrameSinks, that
guarantees the GL context/surface are destroyed. This eliminates a big
source of GPU hangs/crashes with --enable-features=VizDisplayCompositor.

Bug: 777594
Cq-Include-Trybots: master.tryserver.chromium.android:android_optional_gpu_tests_rel
Change-Id: I70e9b031e80de233856194f1beaa9377f332c3c0
Reviewed-on: https://chromium-review.googlesource.com/874724
Commit-Queue: kylechar <kylechar@chromium.org>
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Reviewed-by: Ken Rockot <rockot@chromium.org>
Reviewed-by: Antoine Labour <piman@chromium.org>
Cr-Commit-Position: refs/heads/master@{#531889}

CrOS-Libchrome-Original-Commit: 698da7e959bcea6758232ee56a298a52cd74ddd8