Add a static_assert to ban non-ref-counted raw-pointer receiver on base::Bind
base::Bind used to fail if a raw pointer to a non-ref-counted object is
passed as a receiver. Though that is an expected failure, its error
was hard to read. This CL adds a static_assert for the failing case
for better error message.

Change-Id: I7057fbfaa0c24530f3b9e28b4362f4b8637cd1be
Reviewed-on: https://chromium-review.googlesource.com/923661
Commit-Queue: Taiju Tsuiki <tzik@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/master@{#539772}

CrOS-Libchrome-Original-Commit: 4625ac6130a090c971a3e93f021241db635b0c25