mojo_unittests: Decrease cost of MessageTest.ExtendMessagePayloadLarge
This test is timing out on Windows debug, where even on very fast
devices it can take 35+ seconds to run.

The runtime grows linearly with the size of the test message used, so
this cuts that size to one tenth (512 kB instead of 5 MB).

TBR=jcivelli@chromium.org

Bug: 819046
Change-Id: Ib4598981856566dc8ae54de1060840b52e30b959
Reviewed-on: https://chromium-review.googlesource.com/954045
Reviewed-by: Ken Rockot <rockot@chromium.org>
Commit-Queue: Ken Rockot <rockot@chromium.org>
Cr-Commit-Position: refs/heads/master@{#541704}

CrOS-Libchrome-Original-Commit: cf150a3d82e40aedf2190b1e550f8935a8305cdf