Delay Channel::OnError() in case of kDisconnected during Write().
Write() operations to a Channel can fail due to the peer having closed
it, while there are still messages waiting to be read from it. We must
therefore defer notifying the caller of the Channel::Error until we
observe end-of-stream via a readable notification, otherwise those
messages may be dropped (depending on whether the posted OnError task
is processed before or after a pending Channel-readable event).

Bug: 816620
Change-Id: I75bd34a48edf4022809d27ce49f9cfba7a5d4daf
Reviewed-on: https://chromium-review.googlesource.com/956932
Commit-Queue: Wez <wez@chromium.org>
Reviewed-by: Ken Rockot <rockot@chromium.org>
Cr-Commit-Position: refs/heads/master@{#542634}

CrOS-Libchrome-Original-Commit: e960518251867845ed23bf2ab8f042a77afe1885