Speed up module dumps on macOS.
The out of process heap profiler only requires modules to be dumped, not every
single virtual memory region. This CL modifies the interface to support this,
and updates the macOS implementation to be much faster. The other OSes currently
still perform full address space dumps - this can be changed in the future if it
proves to be a performance issue.

Bug: 826913
Change-Id: Ibaa03d3dfae71a2b6d090aed0ab2ffdc4bb090a9
Reviewed-on: https://chromium-review.googlesource.com/986680
Commit-Queue: Erik Chen <erikchen@chromium.org>
Reviewed-by: Primiano Tucci <primiano@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/master@{#548731}

CrOS-Libchrome-Original-Commit: ec6a7a25ba71c1aec895612aceb449fd4d59459d