Coalesce types related to POSIX IO MessagePumps' WatchFileDescriptor impls
WatchableIOMessagePump : new common interface inherited by all POSIX IO
MessagePumps.

Coalesced names for the watcher/controller so they all use the same
(FdWatcher/FdWatchController) names.

Abstracting these was tricky because the API is such that it requires the user
to know the full type (i.e. as it owns the FdWatchController).
base::FileDescriptorWatcher is a cleaner API but isn't yet ready to support use cases
that require the highest throughput (nor is that in scope of this CL).

A fallout of that is that: WatchableIOMessagePump::WatchFileDescriptor couldn't be
made a real virtual method as it requires using the full type of the FdWatchController.
Since this was already implemented by copying implementations, it was here
merely made into a contract-by-comment for subclasses of WatchableIOMessagePump...

Note to reviewer: reviewing/landing this CL establishes a contract that
I will follow it up with a TBRed mass search-and-replace in the codebase
as follows.
  Win:
    MessageLoopForIO::IOHandler => MessagePumpForIO::IOHandler
    MessageLoopForIO::IOContext => MessagePumpForIO::IOContext
  Posix:
    MessageLoopForIO::Watcher => MessagePumpForIO::FdWatcher
    MessageLoopForIO::FileDescriptorWatcher => MessagePumpForIO::FdWatchController
    MessageLoopForIO::Mode => MessagePumpForIO::Mode
    MessageLoopForIO::WATCH_READ => MessagePumpForIO::WATCH_READ
    MessageLoopForIO::WATCH_WRITE => MessagePumpForIO::WATCH_WRITE
    MessageLoopForIO::WATCH_READ_WRITE => MessagePumpForIO::WATCH_READ_WRITE
  Fuchsia:
    MessageLoopForIO::ZxHandleWatchController = MessagePumpForIO::ZxHandleWatchController;
    MessageLoopForIO::ZxHandleWatcher = MessagePumpForIO::ZxHandleWatcher;

(adding message_pump_for_io.h where needed)

This will allow the MessageLoop interface to no longer need to include
message_pump_for_io.h for itself (and all of its users) and will make
it much easier to split its interface in two (ref.
upcoming base::MessageLoopCurrent @
https://chromium-review.googlesource.com/c/chromium/src/+/957760/9 which
currently trips on those types and ifdefs).

Bug: 825327
Change-Id: I095a966e9f12b1822e1bc77f387f4c2a2772b3f0
Reviewed-on: https://chromium-review.googlesource.com/999795
Commit-Queue: Gabriel Charette <gab@chromium.org>
Reviewed-by: François Doray <fdoray@chromium.org>
Cr-Commit-Position: refs/heads/master@{#549323}

CrOS-Libchrome-Original-Commit: 60a2b0080f0157d5c4d511b74add7a365b5b996c