Replace Verdana with Arimo GCS font
Verdana is NOT metrically-compatible with Arimo, but it is chosen as the
replacement since they are both from the Sans family.  In all of the layout test
deltas, it appears the choice of Verdana was purely coincidental, so it should
be safe to switch to Arimo.

BUG=787020

Change-Id: Id8c82888b2e844e35200d20532b88d87066792bc
Reviewed-on: https://chromium-review.googlesource.com/1006221
Reviewed-by: Michael Wasserman <msw@chromium.org>
Reviewed-by: Dirk Pranke <dpranke@chromium.org>
Commit-Queue: Thomas Anderson <thomasanderson@chromium.org>
Cr-Commit-Position: refs/heads/master@{#550044}

CrOS-Libchrome-Original-Commit: 5277e7971b54830db265ae6546cb6c75c64de8ee