Migrate MessageLoop::current() to MessageLoopCurrent in /components/policy
These changes were scripted as part of the effort to restrict exposure of the
raw MessageLoop*.

Note that many static methods exposed on MessageLoopCurrent are also deprecated
so it is very well possible that this is migrating to an already deprecated
method.

The goal of this pass is to reduce usage of MessageLoop::current() (and
ultimately make it impossible to obtain the MessageLoop* statically).

As such I will not edit this CL unless the script did something logically wrong.
I defer to owners to fix highlighted usage of already deprecated APIs.

Possible script screw ups / things to look out for in this review:
 - Storing MessageLoopCurrent in a MessageLoop* variable or comparing it against
   one (I will go over that in a separate pass).

Includes should have been stripped if that was the last usage of message_loop.h
in that file.

Please CQ if LGTY

This CL was uploaded by git cl split.

R=pmarko@chromium.org

Bug: 825327
Change-Id: I78406e5535645c6a03c283fe2e22a9f801e7efdd
Reviewed-on: https://chromium-review.googlesource.com/1024447
Commit-Queue: Pavol Marko <pmarko@chromium.org>
Reviewed-by: Pavol Marko <pmarko@chromium.org>
Cr-Commit-Position: refs/heads/master@{#553025}

CrOS-Libchrome-Original-Commit: cf03e96fe6ff1aab55b91d1e98331cf5c2d21ed1