Instantiate test class for parameterized tests at the same time as for non-parameterized tests.
ParameterizedRunnerDelegateFactory used to create the test class when creating
the runner, passing the instance in so that createTest() would just return it.
This meant that all test rules would be created at that time as well, which in
the case of RenderTestRule is before the command line is instantiated, so it
would not have the right value of the render test output directory.
(In addition, it defaults to using the test data directory, so writing failure
files would suceed, but the test runner wouldn't find them when pulling from
the device.)

With this change, the test class is created in the createTest() method to avoid
this and other subtle differences between parameterized and non-parameterized
tests.

Bug: 834712
Change-Id: I49bec0e6dbbf5aebe92f29a0de65c9b60729646a
Reviewed-on: https://chromium-review.googlesource.com/1023854
Reviewed-by: Yaron Friedman <yfriedman@chromium.org>
Commit-Queue: Bernhard Bauer <bauerb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#553043}

CrOS-Libchrome-Original-Commit: ec953303fb739c2c8ce4953ea48d79e7ca95ae02