Mark libchrome as recovery_available for update_engine_sideload
update_engine_sideload builds onto the recovery partition, which means
libchrome needs to be marked as recovery_available, so that it can be built
again with recovery-specific dependencies.

Some modules build for recovery differently -- omitting dependencies,
removing source files, different cflags, etc. Because of this, the builds
for the recovery partition are separate from the builds for the system or
vendor partitions.

Test: build update_engine_sideload
Change-Id: Id31a7170e6c9fcff65e3af4eb04440c14cd5c8fa