CHROMIUM: crypto: Add ASN.1 functions for OpenSSL 1.1
In OpenSSL 1.1, ASN1_STRING_data() is deprecated in favor of the new
ASN_STRING_get0_data(). We can't use the former with 1.1 because using
deprecated symbols means errors during the build, and we can't use
the latter before 1.1 because it doesn't exist. This CL provides an
implementation of ASN_STRING_get0_data() pre-1.1.

BUG=chromium:1024391
TEST=Manual, tested through CL:1914779 which depends on this change

Change-Id: Iceab049a25c995099d3d7291a69c2c04d7117230