Decrease the disabled feature's log level
internal::ScopedBlockingCallWithBaseSyncPrimitives is used
for detect the potential deadlock. But right now I do not
have a solution for the detail implementation because no more
time for me.

And it needs to decrease the log level for it, because it's just
useful information for us but error message.

Have fun!